@@ -1,8 +1,8 @@
 {
   "name": "katex",
-  "version": "0.7.0-pre",
+  "version": "0.8.4-pre",
   "description": "Fast math typesetting for the web.",
-  "main": "katex.js",
+  "main": "dist/katex.js",
   "repository": {
     "type": "git",
     "url": "git://github.com/Khan/KaTeX.git"
@@ -11,31 +11,66 @@
     "katex.js",
     "cli.js",
     "src/",
+    "contrib/",
     "dist/"
   ],
   "license": "MIT",
   "devDependencies": {
-    "browserify": "^10.2.4",
-    "clean-css": "~2.2.15",
-    "eslint": "^1.10.2",
-    "express": "~3.3.3",
-    "glob": "^5.0.15",
-    "jasmine": "^2.3.2",
-    "jasmine-core": "^2.3.4",
+    "babel-eslint": "^7.2.0",
+    "babel-jest": "^20.0.3",
+    "babel-plugin-transform-class-properties": "^6.23.0",
+    "babel-plugin-transform-runtime": "^6.15.0",
+    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-flow": "^6.23.0",
+    "babel-register": "^6.24.0",
+    "babelify": "^7.3.0",
+    "browserify": "^13.3.0",
+    "check-dependencies": "^1.1.0",
+    "clean-css": "^3.4.23",
+    "eslint": "^4.6.1",
+    "eslint-plugin-flowtype": "^2.35.1",
+    "express": "^4.14.0",
+    "flow-bin": "^0.54.0",
+    "glob": "^7.1.1",
+    "jest": "^20.0.4",
+    "jest-serializer-html": "^4.0.0",
     "js-yaml": "^3.3.1",
-    "jspngopt": "^0.1.0",
+    "jspngopt": "^0.2.0",
     "less": "~2.7.1",
+    "morgan": "^1.7.0",
     "nomnom": "^1.8.1",
-    "pako": "0.2.7",
-    "selenium-webdriver": "^2.46.1",
-    "uglify-js": "~2.4.15"
+    "object-assign": "^4.1.0",
+    "pako": "1.0.4",
+    "pre-commit": "^1.2.2",
+    "selenium-webdriver": "^2.48.2",
+    "sri-toolbox": "^0.2.0",
+    "uglify-js": "~2.7.5"
   },
   "bin": "cli.js",
   "scripts": {
-    "test": "make lint test",
-    "prepublish": "make dist"
+    "lint": "eslint katex.js server.js cli.js src test contrib dockers",
+    "flow": "flow",
+    "jest": "jest",
+    "coverage": "jest --coverage",
+    "test": "check-dependencies && npm run lint && npm run flow && npm run jest",
+    "start": "check-dependencies && node server.js",
+    "prepublishOnly": "make NIS= dist"
   },
+  "pre-commit": [
+    "lint"
+  ],
   "dependencies": {
-    "match-at": "^0.1.0"
+    "match-at": "^0.1.1"
+  },
+  "jest": {
+    "snapshotSerializers": [
+      "jest-serializer-html"
+    ],
+    "testMatch": [
+      "**/test/*-spec.js"
+    ],
+    "transform": {
+      "^.+\\.js$": "babel-jest"
+    }
   }
 }
