@@ -1,33 +1,31 @@
-var buildHTML = require("./buildHTML");
-// var buildMathML = require("./buildMathML");
-var buildCommon = require("./buildCommon");
-var Options = require("./Options");
-var Settings = require("./Settings");
-var Style = require("./Style");
+import buildHTML from "./buildHTML";
+import buildMathML from "./buildMathML";
+import { makeSpan } from "./buildCommon";
+import Options from "./Options";
+import Settings from "./Settings";
+import Style from "./Style";
 
-var makeSpan = buildCommon.makeSpan;
-
-var buildTree = function(tree, expression, settings) {
+const buildTree = function(tree, expression, settings) {
     settings = settings || new Settings({});
 
-    var startStyle = Style.TEXT;
+    let startStyle = Style.TEXT;
     if (settings.displayMode) {
         startStyle = Style.DISPLAY;
     }
 
     // Setup the default options
-    var options = new Options({
+    const options = new Options({
         style: startStyle,
-        size: "size5",
+        maxSize: settings.maxSize,
     });
 
     // `buildHTML` sometimes messes with the parse tree (like turning bins ->
     // ords), so we build the MathML version first.
-    // var mathMLNode = buildMathML(tree, expression, options);
-    var htmlNode = buildHTML(tree, options);
+    const mathMLNode = buildMathML(tree, expression, options);
+    const htmlNode = buildHTML(tree, options);
 
-    var katexNode = makeSpan(["katex"], [
-        htmlNode,
+    const katexNode = makeSpan(["katex"], [
+        mathMLNode, htmlNode,
     ]);
 
     if (settings.displayMode) {
