@@ -1,112 +1,38 @@
-var utils = require("./utils");
-var ParseError = require("./ParseError");
+// @flow
+/** Include this to ensure that all functions are defined. */
+import utils from "./utils";
+import ParseError from "./ParseError";
+import ParseNode from "./ParseNode";
+import {
+    default as _defineFunction,
+    ordargument,
+    _functions,
+} from "./defineFunction";
 
-/* This file contains a list of functions that we parse, identified by
- * the calls to defineFunction.
- *
- * The first argument to defineFunction is a single name or a list of names.
- * All functions named in such a list will share a single implementation.
- *
- * Each declared function can have associated properties, which
- * include the following:
- *
- *  - numArgs: The number of arguments the function takes.
- *             If this is the only property, it can be passed as a number
- *             instead of an element of a properties object.
- *  - argTypes: (optional) An array corresponding to each argument of the
- *              function, giving the type of argument that should be parsed. Its
- *              length should be equal to `numArgs + numOptionalArgs`. Valid
- *              types:
- *               - "size": A size-like thing, such as "1em" or "5ex"
- *               - "color": An html color, like "#abc" or "blue"
- *               - "original": The same type as the environment that the
- *                             function being parsed is in (e.g. used for the
- *                             bodies of functions like \color where the first
- *                             argument is special and the second argument is
- *                             parsed normally)
- *              Other possible types (probably shouldn't be used)
- *               - "text": Text-like (e.g. \text)
- *               - "math": Normal math
- *              If undefined, this will be treated as an appropriate length
- *              array of "original" strings
- *  - greediness: (optional) The greediness of the function to use ungrouped
- *                arguments.
- *
- *                E.g. if you have an expression
- *                  \sqrt \frac 1 2
- *                since \frac has greediness=2 vs \sqrt's greediness=1, \frac
- *                will use the two arguments '1' and '2' as its two arguments,
- *                then that whole function will be used as the argument to
- *                \sqrt. On the other hand, the expressions
- *                  \frac \frac 1 2 3
- *                and
- *                  \frac \sqrt 1 2
- *                will fail because \frac and \frac have equal greediness
- *                and \sqrt has a lower greediness than \frac respectively. To
- *                make these parse, we would have to change them to:
- *                  \frac {\frac 1 2} 3
- *                and
- *                  \frac {\sqrt 1} 2
- *
- *                The default value is `1`
- *  - allowedInText: (optional) Whether or not the function is allowed inside
- *                   text mode (default false)
- *  - numOptionalArgs: (optional) The number of optional arguments the function
- *                     should parse. If the optional arguments aren't found,
- *                     `null` will be passed to the handler in their place.
- *                     (default 0)
- *  - infix: (optional) Must be true if the function is an infix operator.
- *
- * The last argument is that implementation, the handler for the function(s).
- * It is called to handle these functions and their arguments.
- * It receives two arguments:
- *  - context contains information and references provided by the parser
- *  - args is an array of arguments obtained from TeX input
- * The context contains the following properties:
- *  - funcName: the text (i.e. name) of the function, including \
- *  - parser: the parser object
- *  - lexer: the lexer object
- *  - positions: the positions in the overall string of the function
- *               and the arguments.
- * The latter three should only be used to produce error messages.
- *
- * The function should return an object with the following keys:
- *  - type: The type of element that this is. This is then used in
- *          buildHTML/buildMathML to determine which function
- *          should be called to build this node into a DOM node
- * Any other data can be added to the object, which will be passed
- * in to the function in buildHTML/buildMathML as `group.value`.
- */
+import type {FunctionPropSpec, FunctionHandler} from "./defineFunction" ;
 
-function defineFunction(names, props, handler) {
-    if (typeof names === "string") {
-        names = [names];
-    }
-    if (typeof props === "number") {
-        props = { numArgs: props };
-    }
-    // Set default values of functions
-    var data = {
-        numArgs: props.numArgs,
-        argTypes: props.argTypes,
-        greediness: (props.greediness === undefined) ? 1 : props.greediness,
-        allowedInText: !!props.allowedInText,
-        numOptionalArgs: props.numOptionalArgs || 0,
-        infix: !!props.infix,
-        handler: handler,
-    };
-    for (var i = 0; i < names.length; ++i) {
-        module.exports[names[i]] = data;
-    }
-}
+// WARNING: New functions should be added to src/functions and imported here.
+
+const functions = _functions;
+export default functions;
+
+// Define a convenience function that mimcs the old semantics of defineFunction
+// to support existing code so that we can migrate it a little bit at a time.
+const defineFunction = function(
+    names: string[],
+    props: FunctionPropSpec,
+    handler: ?FunctionHandler, // null only if handled in parser
+) {
+    _defineFunction({names, props, handler});
+};
 
 // A normal square root
-defineFunction("\\sqrt", {
+defineFunction(["\\sqrt"], {
     numArgs: 1,
     numOptionalArgs: 1,
 }, function(context, args) {
-    var index = args[0];
-    var body = args[1];
+    const index = args[0];
+    const body = args[1];
     return {
         type: "sqrt",
         body: body,
@@ -114,26 +40,27 @@ defineFunction("\\sqrt", {
     };
 });
 
-// Some non-mathy text
-defineFunction("\\text", {
+// Non-mathy text, possibly in a font
+const textFunctionFonts = {
+    "\\text": undefined, "\\textrm": "mathrm", "\\textsf": "mathsf",
+    "\\texttt": "mathtt", "\\textnormal": "mathrm", "\\textbf": "mathbf",
+    "\\textit": "textit",
+};
+
+defineFunction([
+    "\\text", "\\textrm", "\\textsf", "\\texttt", "\\textnormal",
+    "\\textbf", "\\textit",
+], {
     numArgs: 1,
     argTypes: ["text"],
     greediness: 2,
+    allowedInText: true,
 }, function(context, args) {
-    var body = args[0];
-    // Since the corresponding buildHTML/buildMathML function expects a
-    // list of elements, we normalize for different kinds of arguments
-    // TODO(emily): maybe this should be done somewhere else
-    var inner;
-    if (body.type === "ordgroup") {
-        inner = body.value;
-    } else {
-        inner = [body];
-    }
-
+    const body = args[0];
     return {
         type: "text",
-        body: inner,
+        body: ordargument(body),
+        font: textFunctionFonts[context.funcName],
     };
 });
 
@@ -259,34 +186,70 @@ defineFunction("\\newline", {
 // });
 
 // A two-argument custom color
-defineFunction("\\color", {
+defineFunction(["\\textcolor"], {
     numArgs: 2,
     allowedInText: true,
     greediness: 3,
     argTypes: ["color", "original"],
 }, function(context, args) {
-    var color = args[0];
-    var body = args[1];
-    // Normalize the different kinds of bodies (see \text above)
-    var inner;
-    if (body.type === "ordgroup") {
-        inner = body.value;
-    } else {
-        inner = [body];
-    }
-
+    const color = args[0];
+    const body = args[1];
     return {
         type: "color",
         color: color.value,
-        value: inner,
+        value: ordargument(body),
+    };
+});
+
+// \color is handled in Parser.js's parseImplicitGroup
+defineFunction(["\\color"], {
+    numArgs: 1,
+    allowedInText: true,
+    greediness: 3,
+    argTypes: ["color"],
+}, null);
+
+// colorbox
+defineFunction(["\\colorbox"], {
+    numArgs: 2,
+    allowedInText: true,
+    greediness: 3,
+    argTypes: ["color", "text"],
+}, function(context, args) {
+    const color = args[0];
+    const body = args[1];
+    return {
+        type: "enclose",
+        label: context.funcName,
+        backgroundColor: color,
+        body: body,
+    };
+});
+
+// fcolorbox
+defineFunction(["\\fcolorbox"], {
+    numArgs: 3,
+    allowedInText: true,
+    greediness: 3,
+    argTypes: ["color", "color", "text"],
+}, function(context, args) {
+    const borderColor = args[0];
+    const backgroundColor = args[1];
+    const body = args[2];
+    return {
+        type: "enclose",
+        label: context.funcName,
+        backgroundColor: backgroundColor,
+        borderColor: borderColor,
+        body: body,
     };
 });
 
 // An overline
-defineFunction("\\overline", {
+defineFunction(["\\overline"], {
     numArgs: 1,
 }, function(context, args) {
-    var body = args[0];
+    const body = args[0];
     return {
         type: "overline",
         body: body,
@@ -294,10 +257,10 @@ defineFunction("\\overline", {
 });
 
 // An underline
-defineFunction("\\underline", {
+defineFunction(["\\underline"], {
     numArgs: 1,
 }, function(context, args) {
-    var body = args[0];
+    const body = args[0];
     return {
         type: "underline",
         body: body,
@@ -305,14 +268,14 @@ defineFunction("\\underline", {
 });
 
 // A box of the width and height
-defineFunction("\\rule", {
+defineFunction(["\\rule"], {
     numArgs: 2,
     numOptionalArgs: 1,
     argTypes: ["size", "size", "size"],
 }, function(context, args) {
-    var shift = args[0];
-    var width = args[1];
-    var height = args[2];
+    const shift = args[0];
+    const width = args[1];
+    const height = args[2];
     return {
         type: "rule",
         shift: shift && shift.value,
@@ -321,7 +284,9 @@ defineFunction("\\rule", {
     };
 });
 
-defineFunction("\\kern", {
+// TODO: In TeX, \mkern only accepts mu-units, and \kern does not accept
+// mu-units. In current KaTeX we relax this; both commands accept any unit.
+defineFunction(["\\kern", "\\mkern"], {
     numArgs: 1,
     argTypes: ["size"],
 }, function(context, args) {
@@ -332,7 +297,7 @@ defineFunction("\\kern", {
 });
 
 // A KaTeX logo
-defineFunction("\\KaTeX", {
+defineFunction(["\\KaTeX"], {
     numArgs: 0,
 }, function(context) {
     return {
@@ -340,59 +305,74 @@ defineFunction("\\KaTeX", {
     };
 });
 
-defineFunction("\\phantom", {
+import "./functions/phantom";
+
+// Math class commands except \mathop
+defineFunction([
+    "\\mathord", "\\mathbin", "\\mathrel", "\\mathopen",
+    "\\mathclose", "\\mathpunct", "\\mathinner",
+], {
     numArgs: 1,
 }, function(context, args) {
-    var body = args[0];
-    var inner;
-    if (body.type === "ordgroup") {
-        inner = body.value;
-    } else {
-        inner = [body];
-    }
-
+    const body = args[0];
     return {
-        type: "phantom",
-        value: inner,
+        type: "mclass",
+        mclass: "m" + context.funcName.substr(5),
+        value: ordargument(body),
     };
 });
 
-// Extra data needed for the delimiter handler down below
-var delimiterSizes = {
-    "\\bigl" : {type: "open",    size: 1},
-    "\\Bigl" : {type: "open",    size: 2},
-    "\\biggl": {type: "open",    size: 3},
-    "\\Biggl": {type: "open",    size: 4},
-    "\\bigr" : {type: "close",   size: 1},
-    "\\Bigr" : {type: "close",   size: 2},
-    "\\biggr": {type: "close",   size: 3},
-    "\\Biggr": {type: "close",   size: 4},
-    "\\bigm" : {type: "rel",     size: 1},
-    "\\Bigm" : {type: "rel",     size: 2},
-    "\\biggm": {type: "rel",     size: 3},
-    "\\Biggm": {type: "rel",     size: 4},
-    "\\big"  : {type: "textord", size: 1},
-    "\\Big"  : {type: "textord", size: 2},
-    "\\bigg" : {type: "textord", size: 3},
-    "\\Bigg" : {type: "textord", size: 4},
-};
+// Build a relation by placing one symbol on top of another
+defineFunction(["\\stackrel"], {
+    numArgs: 2,
+}, function(context, args) {
+    const top = args[0];
+    const bottom = args[1];
 
-var delimiters = [
-    "(", ")", "[", "\\lbrack", "]", "\\rbrack",
-    "\\{", "\\lbrace", "\\}", "\\rbrace",
-    "\\lfloor", "\\rfloor", "\\lceil", "\\rceil",
-    "<", ">", "\\langle", "\\rangle", "\\lt", "\\gt",
-    "\\lvert", "\\rvert", "\\lVert", "\\rVert",
-    "\\lgroup", "\\rgroup", "\\lmoustache", "\\rmoustache",
-    "/", "\\backslash",
-    "|", "\\vert", "\\|", "\\Vert",
-    "\\uparrow", "\\Uparrow",
-    "\\downarrow", "\\Downarrow",
-    "\\updownarrow", "\\Updownarrow",
-    ".",
-];
+    const bottomop = new ParseNode("op", {
+        type: "op",
+        limits: true,
+        alwaysHandleSupSub: true,
+        symbol: false,
+        value: ordargument(bottom),
+    }, bottom.mode);
 
-var fontAliases = {
+    const supsub = new ParseNode("supsub", {
+        base: bottomop,
+        sup: top,
+        sub: null,
+    }, top.mode);
+
+    return {
+        type: "mclass",
+        mclass: "mrel",
+        value: [supsub],
+    };
+});
+
+// \mod-type functions
+defineFunction(["\\bmod"], {
+    numArgs: 0,
+}, function(context, args) {
+    return {
+        type: "mod",
+        modType: "bmod",
+        value: null,
+    };
+});
+
+defineFunction(["\\pod", "\\pmod", "\\mod"], {
+    numArgs: 1,
+}, function(context, args) {
+    const body = args[0];
+    return {
+        type: "mod",
+        modType: context.funcName.substr(1),
+        value: ordargument(body),
+    };
+});
+
+const fontAliases = {
     "\\Bbb": "\\mathbb",
     "\\bold": "\\mathbf",
     "\\frak": "\\mathfrak",
@@ -418,18 +398,11 @@ defineFunction([
     allowedInText: true,
     greediness: 3,
 }, function(context, args) {
-    var body = args[0];
-    var atoms;
-    if (body.type === "ordgroup") {
-        atoms = body.value;
-    } else {
-        atoms = [body];
-    }
-
+    const body = args[0];
     return {
         type: "color",
         color: "katex-" + context.funcName.slice(1),
-        value: atoms,
+        value: ordargument(body),
     };
 });
 
@@ -439,10 +412,11 @@ defineFunction([
 
 // No limits, not symbols
 defineFunction([
-    "\\arcsin", "\\arccos", "\\arctan", "\\arg", "\\cos", "\\cosh",
-    "\\cot", "\\coth", "\\csc", "\\deg", "\\dim", "\\exp", "\\hom",
-    "\\ker", "\\lg", "\\ln", "\\log", "\\sec", "\\sin", "\\sinh",
-    "\\tan", "\\tanh",
+    "\\arcsin", "\\arccos", "\\arctan", "\\arctg", "\\arcctg",
+    "\\arg", "\\ch", "\\cos", "\\cosec", "\\cosh", "\\cot", "\\cotg",
+    "\\coth", "\\csc", "\\ctg", "\\cth", "\\deg", "\\dim", "\\exp",
+    "\\hom", "\\ker", "\\lg", "\\ln", "\\log", "\\sec", "\\sin",
+    "\\sinh", "\\sh", "\\tan", "\\tanh", "\\tg", "\\th",
 ], {
     numArgs: 0,
 }, function(context) {
@@ -499,26 +473,45 @@ defineFunction([
     };
 });
 
+// \mathop class command
+defineFunction(["\\mathop"], {
+    numArgs: 1,
+}, function(context, args) {
+    const body = args[0];
+    return {
+        type: "op",
+        limits: false,
+        symbol: false,
+        value: ordargument(body),
+    };
+});
+
+import "./functions/operators";
+
 // Fractions
 defineFunction([
     "\\dfrac", "\\frac", "\\tfrac",
     "\\dbinom", "\\binom", "\\tbinom",
+    "\\\\atopfrac", // can’t be entered directly
 ], {
     numArgs: 2,
     greediness: 2,
 }, function(context, args) {
-    var numer = args[0];
-    var denom = args[1];
-    var hasBarLine;
-    var leftDelim = null;
-    var rightDelim = null;
-    var size = "auto";
+    const numer = args[0];
+    const denom = args[1];
+    let hasBarLine;
+    let leftDelim = null;
+    let rightDelim = null;
+    let size = "auto";
 
     switch (context.funcName) {
         case "\\dfrac":
         case "\\frac":
         case "\\tfrac":
             hasBarLine = true;
+            break;
+        case "\\\\atopfrac":
+            hasBarLine = false;
             break;
         case "\\dbinom":
         case "\\binom":
@@ -553,64 +546,78 @@ defineFunction([
     };
 });
 
-// Left and right overlap functions
-defineFunction(["\\llap", "\\rlap"], {
+// Horizontal overlap functions
+defineFunction(["\\mathllap", "\\mathrlap", "\\mathclap"], {
     numArgs: 1,
     allowedInText: true,
 }, function(context, args) {
-    var body = args[0];
+    const body = args[0];
     return {
-        type: context.funcName.slice(1),
+        type: "lap",
+        alignment: context.funcName.slice(5),
         body: body,
     };
 });
 
-// Delimiter functions
-defineFunction([
-    "\\bigl", "\\Bigl", "\\biggl", "\\Biggl",
-    "\\bigr", "\\Bigr", "\\biggr", "\\Biggr",
-    "\\bigm", "\\Bigm", "\\biggm", "\\Biggm",
-    "\\big",  "\\Big",  "\\bigg",  "\\Bigg",
-    "\\left", "\\right",
-], {
+// smash, with optional [tb], as in AMS
+defineFunction(["\\smash"], {
     numArgs: 1,
+    numOptionalArgs: 1,
+    allowedInText: true,
 }, function(context, args) {
-    var delim = args[0];
-    if (!utils.contains(delimiters, delim.value)) {
-        throw new ParseError(
-            "Invalid delimiter: '" + delim.value + "' after '" +
-            context.funcName + "'", delim);
+    let smashHeight = false;
+    let smashDepth = false;
+    const tbArg = args[0];
+    if (tbArg) {
+        // Optional [tb] argument is engaged.
+        // ref: amsmath: \renewcommand{\smash}[1][tb]{%
+        //               def\mb@t{\ht}\def\mb@b{\dp}\def\mb@tb{\ht\z@\z@\dp}%
+        let letter = "";
+        for (let i = 0; i < tbArg.value.length; ++i) {
+            letter = tbArg.value[i].value;
+            if (letter === "t") {
+                smashHeight = true;
+            } else if (letter === "b") {
+                smashDepth = true;
+            } else {
+                smashHeight = false;
+                smashDepth = false;
+                break;
+            }
+        }
+    } else {
+        smashHeight = true;
+        smashDepth = true;
     }
 
-    // \left and \right are caught somewhere in Parser.js, which is
-    // why this data doesn't match what is in buildHTML.
-    if (context.funcName === "\\left" || context.funcName === "\\right") {
-        return {
-            type: "leftright",
-            value: delim.value,
-        };
-    } else {
-        return {
-            type: "delimsizing",
-            size: delimiterSizes[context.funcName].size,
-            delimType: delimiterSizes[context.funcName].type,
-            value: delim.value,
-        };
-    }
+    const body = args[1];
+    return {
+        type: "smash",
+        body: body,
+        smashHeight: smashHeight,
+        smashDepth: smashDepth,
+    };
 });
+
+import "./functions/delimsizing";
 
 // Sizing functions (handled in Parser.js explicitly, hence no handler)
 defineFunction([
     "\\tiny", "\\scriptsize", "\\footnotesize", "\\small",
     "\\normalsize", "\\large", "\\Large", "\\LARGE", "\\huge", "\\Huge",
-], 0, null);
+], {numArgs: 0}, null);
 
 // Style changing functions (handled in Parser.js explicitly, hence no
 // handler)
 defineFunction([
     "\\displaystyle", "\\textstyle", "\\scriptstyle",
     "\\scriptscriptstyle",
-], 0, null);
+], {numArgs: 0}, null);
+
+// Old font changing functions
+defineFunction([
+    "\\rm", "\\sf", "\\tt", "\\bf", "\\it", //"\\sl", "\\sc",
+], {numArgs: 0}, null);
 
 defineFunction([
     // styles
@@ -626,8 +633,8 @@ defineFunction([
     numArgs: 1,
     greediness: 2,
 }, function(context, args) {
-    var body = args[0];
-    var func = context.funcName;
+    const body = args[0];
+    let func = context.funcName;
     if (func in fontAliases) {
         func = fontAliases[func];
     }
@@ -642,31 +649,132 @@ defineFunction([
 defineFunction([
     "\\acute", "\\grave", "\\ddot", "\\tilde", "\\bar", "\\breve",
     "\\check", "\\hat", "\\vec", "\\dot",
-    // We don't support expanding accents yet
-    // "\\widetilde", "\\widehat"
+    "\\widehat", "\\widetilde", "\\overrightarrow", "\\overleftarrow",
+    "\\Overrightarrow", "\\overleftrightarrow", "\\overgroup",
+    "\\overlinesegment", "\\overleftharpoon", "\\overrightharpoon",
 ], {
     numArgs: 1,
 }, function(context, args) {
-    var base = args[0];
+    const base = args[0];
+
+    const isStretchy = !utils.contains([
+        "\\acute", "\\grave", "\\ddot", "\\tilde", "\\bar", "\\breve",
+        "\\check", "\\hat", "\\vec", "\\dot",
+    ], context.funcName);
+
+    const isShifty = !isStretchy || utils.contains([
+        "\\widehat", "\\widetilde",
+    ], context.funcName);
+
     return {
         type: "accent",
-        accent: context.funcName,
+        label: context.funcName,
+        isStretchy: isStretchy,
+        isShifty: isShifty,
         base: base,
     };
 });
 
+// Text-mode accents
+defineFunction([
+    "\\'", "\\`", "\\^", "\\~", "\\=", "\\u", "\\.", '\\"',
+    "\\r", "\\H", "\\v",
+], {
+    numArgs: 1,
+    allowedInText: true,
+    allowedInMath: false,
+}, function(context, args) {
+    const base = args[0];
+
+    return {
+        type: "accent",
+        label: context.funcName,
+        isStretchy: false,
+        isShifty: true,
+        base: base,
+    };
+});
+
+// Horizontal stretchy braces
+defineFunction([
+    "\\overbrace", "\\underbrace",
+], {
+    numArgs: 1,
+}, function(context, args) {
+    const base = args[0];
+    return {
+        type: "horizBrace",
+        label: context.funcName,
+        isOver: /^\\over/.test(context.funcName),
+        base: base,
+    };
+});
+
+// Stretchy accents under the body
+defineFunction([
+    "\\underleftarrow", "\\underrightarrow", "\\underleftrightarrow",
+    "\\undergroup", "\\underlinesegment", "\\undertilde",
+], {
+    numArgs: 1,
+}, function(context, args) {
+    const base = args[0];
+    return {
+        type: "accentUnder",
+        label: context.funcName,
+        base: base,
+    };
+});
+
+// Stretchy arrows with an optional argument
+defineFunction([
+    "\\xleftarrow", "\\xrightarrow", "\\xLeftarrow", "\\xRightarrow",
+    "\\xleftrightarrow", "\\xLeftrightarrow", "\\xhookleftarrow",
+    "\\xhookrightarrow", "\\xmapsto", "\\xrightharpoondown",
+    "\\xrightharpoonup", "\\xleftharpoondown", "\\xleftharpoonup",
+    "\\xrightleftharpoons", "\\xleftrightharpoons", "\\xLongequal",
+    "\\xtwoheadrightarrow", "\\xtwoheadleftarrow", "\\xLongequal",
+    "\\xtofrom",
+], {
+    numArgs: 1,
+    numOptionalArgs: 1,
+}, function(context, args) {
+    const below = args[0];
+    const body = args[1];
+    return {
+        type: "xArrow",   // x for extensible
+        label: context.funcName,
+        body: body,
+        below: below,
+    };
+});
+
+// enclose
+defineFunction(["\\cancel", "\\bcancel", "\\xcancel", "\\sout", "\\fbox"], {
+    numArgs: 1,
+}, function(context, args) {
+    const body = args[0];
+    return {
+        type: "enclose",
+        label: context.funcName,
+        body: body,
+    };
+});
+
 // Infix generalized fractions
-defineFunction(["\\over", "\\choose"], {
+defineFunction(["\\over", "\\choose", "\\atop"], {
     numArgs: 0,
     infix: true,
 }, function(context) {
-    var replaceWith;
+    let replaceWith;
     switch (context.funcName) {
         case "\\over":
             replaceWith = "\\frac";
             break;
         case "\\choose":
             replaceWith = "\\binom";
+            break;
+        case "\\atop":
+            replaceWith = "\\\\atopfrac";
             break;
         default:
             throw new Error("Unrecognized infix genfrac command");
@@ -679,34 +787,50 @@ defineFunction(["\\over", "\\choose"], {
 });
 
 // Row breaks for aligned data
-// defineFunction(["\\\\", "\\cr"], {
-//     numArgs: 0,
-//     numOptionalArgs: 1,
-//     argTypes: ["size"],
-// }, function(context, args) {
-//     var size = args[0];
-//     return {
-//         type: "testtag",
-//         size: size,
-//     };
-// });
+defineFunction(["\\\\", "\\cr"], {
+    numArgs: 0,
+    numOptionalArgs: 1,
+    argTypes: ["size"],
+}, function(context, args) {
+    const size = args[0];
+    return {
+        type: "cr",
+        size: size,
+    };
+});
 
 // Environment delimiters
 defineFunction(["\\begin", "\\end"], {
     numArgs: 1,
     argTypes: ["text"],
 }, function(context, args) {
-    var nameGroup = args[0];
+    const nameGroup = args[0];
     if (nameGroup.type !== "ordgroup") {
         throw new ParseError("Invalid environment name", nameGroup);
     }
-    var name = "";
-    for (var i = 0; i < nameGroup.value.length; ++i) {
+    let name = "";
+    for (let i = 0; i < nameGroup.value.length; ++i) {
         name += nameGroup.value[i].value;
     }
     return {
         type: "environment",
         name: name,
         nameGroup: nameGroup,
+    };
+});
+
+// Box manipulation
+defineFunction(["\\raisebox"], {
+    numArgs: 2,
+    argTypes: ["size", "text"],
+    allowedInText: true,
+}, function(context, args) {
+    const amount = args[0];
+    const body = args[1];
+    return {
+        type: "raisebox",
+        dy: amount,
+        body: body,
+        value: ordargument(body),
     };
 });
