@@ -1,32 +1,32 @@
 /* eslint max-len:0 */
 /* global beforeEach: false */
-/* global jasmine: false */
 /* global expect: false */
 /* global it: false */
 /* global describe: false */
 
-var buildMathML = require("../src/buildMathML");
-var buildTree = require("../src/buildTree");
-var katex = require("../katex");
-var ParseError = require("../src/ParseError");
-var parseTree = require("../src/parseTree");
-var Options = require("../src/Options");
-var Settings = require("../src/Settings");
-var Style = require("../src/Style");
+import buildMathML from "../src/buildMathML";
+import buildTree from "../src/buildTree";
+import katex from "../katex";
+import ParseError from "../src/ParseError";
+import parseTree from "../src/parseTree";
+import Options from "../src/Options";
+import Settings from "../src/Settings";
+import Style from "../src/Style";
 
-var defaultSettings = new Settings({});
-var defaultOptions = new Options({
+const defaultSettings = new Settings({});
+const defaultOptions = new Options({
     style: Style.TEXT,
-    size: "size5",
+    size: 5,
+    maxSize: Infinity,
 });
 
-var _getBuilt = function(expr, settings) {
-    var usedSettings = settings ? settings : defaultSettings;
-    var parsedTree = parseTree(expr, usedSettings);
-    var rootNode = buildTree(parsedTree, expr, usedSettings);
+const _getBuilt = function(expr, settings) {
+    const usedSettings = settings ? settings : defaultSettings;
+    const parsedTree = parseTree(expr, usedSettings);
+    const rootNode = buildTree(parsedTree, expr, usedSettings);
 
     // grab the root node of the HTML rendering
-    var builtHTML = rootNode.children[1];
+    const builtHTML = rootNode.children[1];
 
     // Remove the outer .katex and .katex-inner layers
     return builtHTML.children[2].children;
@@ -38,8 +38,8 @@ var _getBuilt = function(expr, settings) {
  * @param settings
  * @returns {Object}
  */
-var getBuilt = function(expr, settings) {
-    var usedSettings = settings ? settings : defaultSettings;
+const getBuilt = function(expr, settings) {
+    const usedSettings = settings ? settings : defaultSettings;
     expect(expr).toBuild(usedSettings);
     return _getBuilt(expr, settings);
 };
@@ -50,14 +50,14 @@ var getBuilt = function(expr, settings) {
  * @param settings
  * @returns {Object}
  */
-var getParsed = function(expr, settings) {
-    var usedSettings = settings ? settings : defaultSettings;
+const getParsed = function(expr, settings) {
+    const usedSettings = settings ? settings : defaultSettings;
 
     expect(expr).toParse(usedSettings);
     return parseTree(expr, usedSettings);
 };
 
-var stripPositions = function(expr) {
+const stripPositions = function(expr) {
     if (typeof expr !== "object" || expr === null) {
         return expr;
     }
@@ -72,7 +72,7 @@ var stripPositions = function(expr) {
     return expr;
 };
 
-var parseAndSetResult = function(expr, result, settings) {
+const parseAndSetResult = function(expr, result, settings) {
     try {
         return parseTree(expr, settings || defaultSettings);
     } catch (e) {
@@ -88,111 +88,99 @@ var parseAndSetResult = function(expr, result, settings) {
 };
 
 beforeEach(function() {
-    jasmine.addMatchers({
+    expect.extend({
+        toParse: function(actual, settings) {
+            const usedSettings = settings ? settings : defaultSettings;
 
-        toParse: function() {
-            return {
-                compare: function(actual, settings) {
-                    var usedSettings = settings ? settings : defaultSettings;
-
-                    var result = {
-                        pass: true,
-                        message: "'" + actual + "' succeeded parsing",
-                    };
-                    parseAndSetResult(actual, result, usedSettings);
-                    return result;
-                },
+            const result = {
+                pass: true,
+                message: "'" + actual + "' succeeded parsing",
             };
+            parseAndSetResult(actual, result, usedSettings);
+            return result;
         },
 
-        toNotParse: function() {
-            return {
-                compare: function(actual, settings) {
-                    var usedSettings = settings ? settings : defaultSettings;
+        toNotParse: function(actual, settings) {
+            const usedSettings = settings ? settings : defaultSettings;
 
-                    var result = {
-                        pass: false,
-                        message: "Expected '" + actual + "' to fail " +
-                            "parsing, but it succeeded",
-                    };
-
-                    try {
-                        parseTree(actual, usedSettings);
-                    } catch (e) {
-                        if (e instanceof ParseError) {
-                            result.pass = true;
-                            result.message = "'" + actual + "' correctly " +
-                                "didn't parse with error: " + e.message;
-                        } else {
-                            result.message = "'" + actual + "' failed " +
-                                "parsing with unknown error: " + e.message;
-                        }
-                    }
-
-                    return result;
-                },
+            const result = {
+                pass: false,
+                message: "Expected '" + actual + "' to fail " +
+                    "parsing, but it succeeded",
             };
+
+            try {
+                parseTree(actual, usedSettings);
+            } catch (e) {
+                if (e instanceof ParseError) {
+                    result.pass = true;
+                    result.message = "'" + actual + "' correctly " +
+                        "didn't parse with error: " + e.message;
+                } else {
+                    result.message = "'" + actual + "' failed " +
+                        "parsing with unknown error: " + e.message;
+                }
+            }
+
+            return result;
         },
 
-        toBuild: function() {
-            return {
-                compare: function(actual, settings) {
-                    var usedSettings = settings ? settings : defaultSettings;
+        toBuild: function(actual, settings) {
+            const usedSettings = settings ? settings : defaultSettings;
 
-                    var result = {
-                        pass: true,
-                        message: "'" + actual + "' succeeded in building",
-                    };
-
-                    expect(actual).toParse(usedSettings);
-
-                    try {
-                        _getBuilt(actual, settings);
-                    } catch (e) {
-                        result.pass = false;
-                        if (e instanceof ParseError) {
-                            result.message = "'" + actual + "' failed to " +
-                                "build with error: " + e.message;
-                        } else {
-                            result.message = "'" + actual + "' failed " +
-                                "building with unknown error: " + e.message;
-                        }
-                    }
-
-                    return result;
-                },
+            const result = {
+                pass: true,
+                message: "'" + actual + "' succeeded in building",
             };
+
+            expect(actual).toParse(usedSettings);
+
+            try {
+                _getBuilt(actual, settings);
+            } catch (e) {
+                result.pass = false;
+                if (e instanceof ParseError) {
+                    result.message = "'" + actual + "' failed to " +
+                        "build with error: " + e.message;
+                } else {
+                    result.message = "'" + actual + "' failed " +
+                        "building with unknown error: " + e.message;
+                }
+            }
+
+            return result;
         },
 
-        toParseLike: function(util, baton) {
-            return {
-                compare: function(actual, expected) {
-                    var result = {
-                        pass: true,
-                        message: "Parse trees of '" + actual +
-                            "' and '" + expected + "' are equivalent",
-                    };
+        toParseLike: function(actual, expected, settings) {
+            const usedSettings = settings ? settings : defaultSettings;
 
-                    var actualTree = parseAndSetResult(actual, result);
-                    if (!actualTree) {
-                        return result;
-                    }
-                    var expectedTree = parseAndSetResult(expected, result);
-                    if (!expectedTree) {
-                        return result;
-                    }
-                    stripPositions(actualTree);
-                    stripPositions(expectedTree);
-                    if (!util.equals(actualTree, expectedTree, baton)) {
-                        result.pass = false;
-                        result.message = "Parse trees of '" + actual +
-                            "' and '" + expected + "' are not equivalent";
-                    }
-                    return result;
-                },
+            const result = {
+                pass: true,
+                message: "Parse trees of '" + actual +
+                    "' and '" + expected + "' are equivalent",
             };
-        },
 
+            const actualTree = parseAndSetResult(actual, result,
+                usedSettings);
+            if (!actualTree) {
+                return result;
+            }
+            const expectedTree = parseAndSetResult(expected, result,
+                usedSettings);
+            if (!expectedTree) {
+                return result;
+            }
+
+            stripPositions(actualTree);
+            stripPositions(expectedTree);
+
+            if (JSON.stringify(actualTree) !== JSON.stringify(expectedTree)) {
+                result.pass = false;
+                result.message = "Parse trees of '" + actual +
+                    "' and '" + expected + "' are not equivalent";
+            }
+            return result;
+        },
     });
 });
 
@@ -202,122 +190,124 @@ describe("A parser", function() {
     });
 
     it("should ignore whitespace", function() {
-        var parseA = stripPositions(getParsed("    x    y    "));
-        var parseB = stripPositions(getParsed("xy"));
+        const parseA = stripPositions(getParsed("    x    y    "));
+        const parseB = stripPositions(getParsed("xy"));
         expect(parseA).toEqual(parseB);
     });
 });
 
 describe("An ord parser", function() {
-    var expression = "1234|/@.\"`abcdefgzABCDEFGZ";
+    const expression = "1234|/@.\"`abcdefgzABCDEFGZ";
 
     it("should not fail", function() {
         expect(expression).toParse();
     });
 
     it("should build a list of ords", function() {
-        var parse = getParsed(expression);
+        const parse = getParsed(expression);
 
         expect(parse).toBeTruthy();
 
-        for (var i = 0; i < parse.length; i++) {
-            var group = parse[i];
+        for (let i = 0; i < parse.length; i++) {
+            const group = parse[i];
             expect(group.type).toMatch("ord");
         }
     });
 
     it("should parse the right number of ords", function() {
-        var parse = getParsed(expression);
+        const parse = getParsed(expression);
 
         expect(parse.length).toBe(expression.length);
     });
 });
 
 describe("A bin parser", function() {
-    var expression = "+-*\\cdot\\pm\\div";
+    const expression = "+-*\\cdot\\pm\\div";
 
     it("should not fail", function() {
         expect(expression).toParse();
     });
 
     it("should build a list of bins", function() {
-        var parse = getParsed(expression);
+        const parse = getParsed(expression);
         expect(parse).toBeTruthy();
 
-        for (var i = 0; i < parse.length; i++) {
-            var group = parse[i];
+        for (let i = 0; i < parse.length; i++) {
+            const group = parse[i];
             expect(group.type).toEqual("bin");
         }
     });
 });
 
 describe("A rel parser", function() {
-    var expression = "=<>\\leq\\geq\\neq\\nleq\\ngeq\\cong";
+    const expression = "=<>\\leq\\geq\\neq\\nleq\\ngeq\\cong";
+    const notExpression = "\\not=\\not<\\not>\\not\\leq\\not\\geq\\not\\in";
 
     it("should not fail", function() {
         expect(expression).toParse();
+        expect(notExpression).toParse();
     });
 
     it("should build a list of rels", function() {
-        var parse = getParsed(expression);
+        const parse = getParsed(expression);
         expect(parse).toBeTruthy();
 
-        for (var i = 0; i < parse.length; i++) {
-            var group = parse[i];
+        for (let i = 0; i < parse.length; i++) {
+            const group = parse[i];
             expect(group.type).toEqual("rel");
         }
     });
 });
 
 describe("A punct parser", function() {
-    var expression = ",;\\colon";
+    const expression = ",;\\colon";
 
     it("should not fail", function() {
         expect(expression).toParse();
     });
 
     it("should build a list of puncts", function() {
-        var parse = getParsed(expression);
+        const parse = getParsed(expression);
         expect(parse).toBeTruthy();
 
-        for (var i = 0; i < parse.length; i++) {
-            var group = parse[i];
+        for (let i = 0; i < parse.length; i++) {
+            const group = parse[i];
             expect(group.type).toEqual("punct");
         }
     });
 });
 
 describe("An open parser", function() {
-    var expression = "([";
+    const expression = "([";
 
     it("should not fail", function() {
         expect(expression).toParse();
     });
 
     it("should build a list of opens", function() {
-        var parse = getParsed(expression);
+        const parse = getParsed(expression);
         expect(parse).toBeTruthy();
 
-        for (var i = 0; i < parse.length; i++) {
-            var group = parse[i];
+        for (let i = 0; i < parse.length; i++) {
+            const group = parse[i];
             expect(group.type).toEqual("open");
         }
     });
 });
 
 describe("A close parser", function() {
-    var expression = ")]?!";
+    const expression = ")]?!";
 
     it("should not fail", function() {
         expect(expression).toParse();
     });
 
     it("should build a list of closes", function() {
-        var parse = getParsed(expression);
+        const parse = getParsed(expression);
         expect(parse).toBeTruthy();
 
-        for (var i = 0; i < parse.length; i++) {
-            var group = parse[i];
+        for (let i = 0; i < parse.length; i++) {
+            const group = parse[i];
             expect(group.type).toEqual("close");
         }
     });
@@ -352,7 +342,7 @@ describe("A subscript and superscript parser", function() {
     });
 
     it("should produce supsubs for superscript", function() {
-        var parse = getParsed("x^2")[0];
+        const parse = getParsed("x^2")[0];
 
         expect(parse.type).toBe("supsub");
         expect(parse.value.base).toBeDefined();
@@ -361,7 +351,7 @@ describe("A subscript and superscript parser", function() {
     });
 
     it("should produce supsubs for subscript", function() {
-        var parse = getParsed("x_3")[0];
+        const parse = getParsed("x_3")[0];
 
         expect(parse.type).toBe("supsub");
         expect(parse.value.base).toBeDefined();
@@ -370,7 +360,7 @@ describe("A subscript and superscript parser", function() {
     });
 
     it("should produce supsubs for ^_", function() {
-        var parse = getParsed("x^2_3")[0];
+        const parse = getParsed("x^2_3")[0];
 
         expect(parse.type).toBe("supsub");
         expect(parse.value.base).toBeDefined();
@@ -379,7 +369,7 @@ describe("A subscript and superscript parser", function() {
     });
 
     it("should produce supsubs for _^", function() {
-        var parse = getParsed("x_3^2")[0];
+        const parse = getParsed("x_3^2")[0];
 
         expect(parse.type).toBe("supsub");
         expect(parse.value.base).toBeDefined();
@@ -388,8 +378,8 @@ describe("A subscript and superscript parser", function() {
     });
 
     it("should produce the same thing regardless of order", function() {
-        var parseA = stripPositions(getParsed("x^2_3"));
-        var parseB = stripPositions(getParsed("x_3^2"));
+        const parseA = stripPositions(getParsed("x^2_3"));
+        const parseB = stripPositions(getParsed("x_3^2"));
 
         expect(parseA).toEqual(parseB);
     });
@@ -470,7 +460,7 @@ describe("A parser with limit controls", function() {
     it("should have the rightmost limit control determine the limits property " +
         "of the preceding op node", function() {
 
-        var parsedInput = getParsed("\\int\\nolimits\\limits_2^2");
+        let parsedInput = getParsed("\\int\\nolimits\\limits_2^2");
         expect(parsedInput[0].value.base.value.limits).toBe(true);
 
         parsedInput = getParsed("\\int\\limits_2\\nolimits^2");
@@ -484,11 +474,11 @@ describe("A group parser", function() {
     });
 
     it("should produce a single ord", function() {
-        var parse = getParsed("{xy}");
+        const parse = getParsed("{xy}");
 
         expect(parse.length).toBe(1);
 
-        var ord = parse[0];
+        const ord = parse[0];
 
         expect(ord.type).toMatch("ord");
         expect(ord.value).toBeTruthy();
@@ -502,32 +492,32 @@ describe("An implicit group parser", function() {
     });
 
     it("should produce a single object", function() {
-        var parse = getParsed("\\Large abc");
+        const parse = getParsed("\\Large abc");
 
         expect(parse.length).toBe(1);
 
-        var sizing = parse[0];
+        const sizing = parse[0];
 
         expect(sizing.type).toEqual("sizing");
         expect(sizing.value).toBeTruthy();
     });
 
     it("should apply only after the function", function() {
-        var parse = getParsed("a \\Large abc");
+        const parse = getParsed("a \\Large abc");
 
         expect(parse.length).toBe(2);
 
-        var sizing = parse[1];
+        const sizing = parse[1];
 
         expect(sizing.type).toEqual("sizing");
         expect(sizing.value.value.length).toBe(3);
     });
 
     it("should stop at the ends of groups", function() {
-        var parse = getParsed("a { b \\Large c } d");
+        const parse = getParsed("a { b \\Large c } d");
 
-        var group = parse[1];
-        var sizing = group.value[1];
+        const group = parse[1];
+        const sizing = group.value[1];
 
         expect(sizing.type).toEqual("sizing");
         expect(sizing.value.value.length).toBe(1);
@@ -571,16 +561,16 @@ describe("A function parser", function() {
 });
 
 describe("A frac parser", function() {
-    var expression = "\\frac{x}{y}";
-    var dfracExpression = "\\dfrac{x}{y}";
-    var tfracExpression = "\\tfrac{x}{y}";
+    const expression = "\\frac{x}{y}";
+    const dfracExpression = "\\dfrac{x}{y}";
+    const tfracExpression = "\\tfrac{x}{y}";
 
     it("should not fail", function() {
         expect(expression).toParse();
     });
 
     it("should produce a frac", function() {
-        var parse = getParsed(expression)[0];
+        const parse = getParsed(expression)[0];
 
         expect(parse.type).toEqual("genfrac");
         expect(parse.value.numer).toBeDefined();
@@ -594,23 +584,32 @@ describe("A frac parser", function() {
     });
 
     it("should parse dfrac and tfrac as fracs", function() {
-        var dfracParse = getParsed(dfracExpression)[0];
+        const dfracParse = getParsed(dfracExpression)[0];
 
         expect(dfracParse.type).toEqual("genfrac");
         expect(dfracParse.value.numer).toBeDefined();
         expect(dfracParse.value.denom).toBeDefined();
 
-        var tfracParse = getParsed(tfracExpression)[0];
+        const tfracParse = getParsed(tfracExpression)[0];
 
         expect(tfracParse.type).toEqual("genfrac");
         expect(tfracParse.value.numer).toBeDefined();
         expect(tfracParse.value.denom).toBeDefined();
     });
+
+    it("should parse atop", function() {
+        const parse = getParsed("x \\atop y")[0];
+
+        expect(parse.type).toEqual("genfrac");
+        expect(parse.value.numer).toBeDefined();
+        expect(parse.value.denom).toBeDefined();
+        expect(parse.value.hasBarLine).toEqual(false);
+    });
 });
 
 describe("An over parser", function() {
-    var simpleOver = "1 \\over x";
-    var complexOver = "1+2i \\over 3+4i";
+    const simpleOver = "1 \\over x";
+    const complexOver = "1+2i \\over 3+4i";
 
     it("should not fail", function() {
         expect(simpleOver).toParse();
@@ -618,7 +617,7 @@ describe("An over parser", function() {
     });
 
     it("should produce a frac", function() {
-        var parse;
+        let parse;
 
         parse = getParsed(simpleOver)[0];
 
@@ -634,38 +633,38 @@ describe("An over parser", function() {
     });
 
     it("should create a numerator from the atoms before \\over", function() {
-        var parse = getParsed(complexOver)[0];
+        const parse = getParsed(complexOver)[0];
 
-        var numer = parse.value.numer;
+        const numer = parse.value.numer;
         expect(numer.value.length).toEqual(4);
     });
 
     it("should create a demonimator from the atoms after \\over", function() {
-        var parse = getParsed(complexOver)[0];
+        const parse = getParsed(complexOver)[0];
 
-        var denom = parse.value.numer;
+        const denom = parse.value.numer;
         expect(denom.value.length).toEqual(4);
     });
 
     it("should handle empty numerators", function() {
-        var emptyNumerator = "\\over x";
-        var parse = getParsed(emptyNumerator)[0];
+        const emptyNumerator = "\\over x";
+        const parse = getParsed(emptyNumerator)[0];
         expect(parse.type).toEqual("genfrac");
         expect(parse.value.numer).toBeDefined();
         expect(parse.value.denom).toBeDefined();
     });
 
     it("should handle empty denominators", function() {
-        var emptyDenominator = "1 \\over";
-        var parse = getParsed(emptyDenominator)[0];
+        const emptyDenominator = "1 \\over";
+        const parse = getParsed(emptyDenominator)[0];
         expect(parse.type).toEqual("genfrac");
         expect(parse.value.numer).toBeDefined();
         expect(parse.value.denom).toBeDefined();
     });
 
     it("should handle \\displaystyle correctly", function() {
-        var displaystyleExpression = "\\displaystyle 1 \\over 2";
-        var parse = getParsed(displaystyleExpression)[0];
+        const displaystyleExpression = "\\displaystyle 1 \\over 2";
+        const parse = getParsed(displaystyleExpression)[0];
         expect(parse.type).toEqual("genfrac");
         expect(parse.value.numer.value[0].type).toEqual("styling");
         expect(parse.value.denom).toBeDefined();
@@ -679,8 +678,8 @@ describe("An over parser", function() {
     });
 
     it("should handle nested factions", function() {
-        var nestedOverExpression = "{1 \\over 2} \\over 3";
-        var parse = getParsed(nestedOverExpression)[0];
+        const nestedOverExpression = "{1 \\over 2} \\over 3";
+        const parse = getParsed(nestedOverExpression)[0];
         expect(parse.type).toEqual("genfrac");
         expect(parse.value.numer.value[0].type).toEqual("genfrac");
         expect(parse.value.numer.value[0].value.numer.value[0].value).toEqual("1");
@@ -690,23 +689,23 @@ describe("An over parser", function() {
     });
 
     it("should fail with multiple overs in the same group", function() {
-        var badMultipleOvers = "1 \\over 2 + 3 \\over 4";
+        const badMultipleOvers = "1 \\over 2 + 3 \\over 4";
         expect(badMultipleOvers).toNotParse();
 
-        var badOverChoose = "1 \\over 2 \\choose 3";
+        const badOverChoose = "1 \\over 2 \\choose 3";
         expect(badOverChoose).toNotParse();
     });
 });
 
 describe("A sizing parser", function() {
-    var sizeExpression = "\\Huge{x}\\small{x}";
+    const sizeExpression = "\\Huge{x}\\small{x}";
 
     it("should not fail", function() {
         expect(sizeExpression).toParse();
     });
 
     it("should produce a sizing node", function() {
-        var parse = getParsed(sizeExpression)[0];
+        const parse = getParsed(sizeExpression)[0];
 
         expect(parse.type).toEqual("sizing");
         expect(parse.value).toBeDefined();
@@ -714,30 +713,31 @@ describe("A sizing parser", function() {
 });
 
 describe("A text parser", function() {
-    var textExpression = "\\text{a b}";
-    var noBraceTextExpression = "\\text x";
-    var nestedTextExpression =
-        "\\text{a {b} \\blue{c} \\color{#fff}{x} \\llap{x}}";
-    var spaceTextExpression = "\\text{  a \\ }";
-    var leadingSpaceTextExpression = "\\text {moo}";
-    var badTextExpression = "\\text{a b%}";
-    var badFunctionExpression = "\\text{\\sqrt{x}}";
-    var mathTokenAfterText = "\\text{sin}^2";
+    const textExpression = "\\text{a b}";
+    const noBraceTextExpression = "\\text x";
+    const nestedTextExpression =
+        "\\text{a {b} \\blue{c} \\textcolor{#fff}{x} \\llap{x}}";
+    const spaceTextExpression = "\\text{  a \\ }";
+    const leadingSpaceTextExpression = "\\text {moo}";
+    const badTextExpression = "\\text{a b%}";
+    const badFunctionExpression = "\\text{\\sqrt{x}}";
+    const mathTokenAfterText = "\\text{sin}^2";
+    const textWithEmbeddedMath = "\\text{graph: $y = mx + b$}";
 
     it("should not fail", function() {
         expect(textExpression).toParse();
     });
 
     it("should produce a text", function() {
-        var parse = getParsed(textExpression)[0];
+        const parse = getParsed(textExpression)[0];
 
         expect(parse.type).toEqual("text");
         expect(parse.value).toBeDefined();
     });
 
     it("should produce textords instead of mathords", function() {
-        var parse = getParsed(textExpression)[0];
-        var group = parse.value.body;
+        const parse = getParsed(textExpression)[0];
+        const group = parse.value.body;
 
         expect(group[0].type).toEqual("textord");
     });
@@ -759,8 +759,8 @@ describe("A text parser", function() {
     });
 
     it("should contract spaces", function() {
-        var parse = getParsed(spaceTextExpression)[0];
-        var group = parse.value.body;
+        const parse = getParsed(spaceTextExpression)[0];
+        const group = parse.value.body;
 
         expect(group[0].type).toEqual("spacing");
         expect(group[1].type).toEqual("textord");
@@ -773,27 +773,40 @@ describe("A text parser", function() {
     });
 
     it("should ignore a space before the text group", function() {
-        var parse = getParsed(leadingSpaceTextExpression)[0];
+        const parse = getParsed(leadingSpaceTextExpression)[0];
         // [m, o, o]
         expect(parse.value.body.length).toBe(3);
         expect(
             parse.value.body.map(function(n) { return n.value; }).join("")
         ).toBe("moo");
     });
+
+    it("should parse math within text group", function() {
+        expect(textWithEmbeddedMath).toParse();
+    });
+
+    it("should omit spaces after commands", function() {
+        expect("\\text{\\textellipsis !}")
+            .toParseLike("\\text{\\textellipsis!}");
+    });
 });
 
 describe("A color parser", function() {
-    var colorExpression = "\\blue{x}";
-    var newColorExpression = "\\redA{x}";
-    var customColorExpression = "\\color{#fA6}{x}";
-    var badCustomColorExpression = "\\color{bad-color}{x}";
+    const colorExpression = "\\blue{x}";
+    const newColorExpression = "\\redA{x}";
+    const customColorExpression1 = "\\textcolor{#fA6}{x}";
+    const customColorExpression2 = "\\textcolor{#fA6fA6}{x}";
+    const badCustomColorExpression1 = "\\textcolor{bad-color}{x}";
+    const badCustomColorExpression2 = "\\textcolor{#fA6f}{x}";
+    const badCustomColorExpression3 = "\\textcolor{#gA6}{x}";
+    const oldColorExpression = "\\color{#fA6}xy";
 
     it("should not fail", function() {
         expect(colorExpression).toParse();
     });
 
     it("should build a color node", function() {
-        var parse = getParsed(colorExpression)[0];
+        const parse = getParsed(colorExpression)[0];
 
         expect(parse.type).toEqual("color");
         expect(parse.value.color).toBeDefined();
@@ -801,17 +814,22 @@ describe("A color parser", function() {
     });
 
     it("should parse a custom color", function() {
-        expect(customColorExpression).toParse();
+        expect(customColorExpression1).toParse();
+        expect(customColorExpression2).toParse();
     });
 
     it("should correctly extract the custom color", function() {
-        var parse = getParsed(customColorExpression)[0];
+        const parse1 = getParsed(customColorExpression1)[0];
+        const parse2 = getParsed(customColorExpression2)[0];
 
-        expect(parse.value.color).toEqual("#fA6");
+        expect(parse1.value.color).toEqual("#fA6");
+        expect(parse2.value.color).toEqual("#fA6fA6");
     });
 
     it("should not parse a bad custom color", function() {
-        expect(badCustomColorExpression).toNotParse();
+        expect(badCustomColorExpression1).toNotParse();
+        expect(badCustomColorExpression2).toNotParse();
+        expect(badCustomColorExpression3).toNotParse();
     });
 
     it("should parse new colors from the branding guide", function() {
@@ -819,16 +837,32 @@ describe("A color parser", function() {
     });
 
     it("should have correct greediness", function() {
-        expect("\\color{red}a").toParse();
-        expect("\\color{red}{\\text{a}}").toParse();
-        expect("\\color{red}\\text{a}").toNotParse();
-        expect("\\color{red}\\frac12").toNotParse();
+        expect("\\textcolor{red}a").toParse();
+        expect("\\textcolor{red}{\\text{a}}").toParse();
+        expect("\\textcolor{red}\\text{a}").toNotParse();
+        expect("\\textcolor{red}\\frac12").toNotParse();
+    });
+
+    it("should use one-argument \\color by default", function() {
+        expect(oldColorExpression).toParseLike("\\textcolor{#fA6}{xy}");
+    });
+
+    it("should use one-argument \\color if requested", function() {
+        expect(oldColorExpression).toParseLike("\\textcolor{#fA6}{xy}", {
+            colorIsTextColor: false,
+        });
+    });
+
+    it("should use two-argument \\color if requested", function() {
+        expect(oldColorExpression).toParseLike("\\textcolor{#fA6}{x}y", {
+            colorIsTextColor: true,
+        });
     });
 });
 
 describe("A tie parser", function() {
-    var mathTie = "a~b";
-    var textTie = "\\text{a~ b}";
+    const mathTie = "a~b";
+    const textTie = "\\text{a~ b}";
 
     it("should parse ties in math mode", function() {
         expect(mathTie).toParse();
@@ -839,30 +873,30 @@ describe("A tie parser", function() {
     });
 
     it("should produce spacing in math mode", function() {
-        var parse = getParsed(mathTie);
+        const parse = getParsed(mathTie);
 
         expect(parse[1].type).toEqual("spacing");
     });
 
     it("should produce spacing in text mode", function() {
-        var text = getParsed(textTie)[0];
-        var parse = text.value.body;
+        const text = getParsed(textTie)[0];
+        const parse = text.value.body;
 
         expect(parse[1].type).toEqual("spacing");
     });
 
     it("should not contract with spaces in text mode", function() {
-        var text = getParsed(textTie)[0];
-        var parse = text.value.body;
+        const text = getParsed(textTie)[0];
+        const parse = text.value.body;
 
         expect(parse[2].type).toEqual("spacing");
     });
 });
 
 describe("A delimiter sizing parser", function() {
-    var normalDelim = "\\bigl |";
-    var notDelim = "\\bigl x";
-    var bigDelim = "\\Biggr \\langle";
+    const normalDelim = "\\bigl |";
+    const notDelim = "\\bigl x";
+    const bigDelim = "\\Biggr \\langle";
 
     it("should parse normal delimiters", function() {
         expect(normalDelim).toParse();
@@ -874,22 +908,22 @@ describe("A delimiter sizing parser", function() {
     });
 
     it("should produce a delimsizing", function() {
-        var parse = getParsed(normalDelim)[0];
+        const parse = getParsed(normalDelim)[0];
 
         expect(parse.type).toEqual("delimsizing");
     });
 
     it("should produce the correct direction delimiter", function() {
-        var leftParse = getParsed(normalDelim)[0];
-        var rightParse = getParsed(bigDelim)[0];
+        const leftParse = getParsed(normalDelim)[0];
+        const rightParse = getParsed(bigDelim)[0];
 
-        expect(leftParse.value.delimType).toEqual("open");
-        expect(rightParse.value.delimType).toEqual("close");
+        expect(leftParse.value.mclass).toEqual("mopen");
+        expect(rightParse.value.mclass).toEqual("mclose");
     });
 
     it("should parse the correct size delimiter", function() {
-        var smallParse = getParsed(normalDelim)[0];
-        var bigParse = getParsed(bigDelim)[0];
+        const smallParse = getParsed(normalDelim)[0];
+        const bigParse = getParsed(bigDelim)[0];
 
         expect(smallParse.value.size).toEqual(1);
         expect(bigParse.value.size).toEqual(4);
@@ -897,26 +931,26 @@ describe("A delimiter sizing parser", function() {
 });
 
 describe("An overline parser", function() {
-    var overline = "\\overline{x}";
+    const overline = "\\overline{x}";
 
     it("should not fail", function() {
         expect(overline).toParse();
     });
 
     it("should produce an overline", function() {
-        var parse = getParsed(overline)[0];
+        const parse = getParsed(overline)[0];
 
         expect(parse.type).toEqual("overline");
     });
 });
 
 describe("A rule parser", function() {
-    var emRule = "\\rule{1em}{2em}";
-    var exRule = "\\rule{1ex}{2em}";
-    var badUnitRule = "\\rule{1px}{2em}";
-    var noNumberRule = "\\rule{1em}{em}";
-    var incompleteRule = "\\rule{1em}";
-    var hardNumberRule = "\\rule{   01.24ex}{2.450   em   }";
+    const emRule = "\\rule{1em}{2em}";
+    const exRule = "\\rule{1ex}{2em}";
+    const badUnitRule = "\\rule{1au}{2em}";
+    const noNumberRule = "\\rule{1em}{em}";
+    const incompleteRule = "\\rule{1em}";
+    const hardNumberRule = "\\rule{   01.24ex}{2.450   em   }";
 
     it("should not fail", function() {
         expect(emRule).toParse();
@@ -934,14 +968,14 @@ describe("A rule parser", function() {
     });
 
     it("should produce a rule", function() {
-        var parse = getParsed(emRule)[0];
+        const parse = getParsed(emRule)[0];
 
         expect(parse.type).toEqual("rule");
     });
 
     it("should list the correct units", function() {
-        var emParse = getParsed(emRule)[0];
-        var exParse = getParsed(exRule)[0];
+        const emParse = getParsed(emRule)[0];
+        const exParse = getParsed(exRule)[0];
 
         expect(emParse.value.width.unit).toEqual("em");
         expect(emParse.value.height.unit).toEqual("em");
@@ -951,14 +985,14 @@ describe("A rule parser", function() {
     });
 
     it("should parse the number correctly", function() {
-        var hardNumberParse = getParsed(hardNumberRule)[0];
+        const hardNumberParse = getParsed(hardNumberRule)[0];
 
         expect(hardNumberParse.value.width.number).toBeCloseTo(1.24);
         expect(hardNumberParse.value.height.number).toBeCloseTo(2.45);
     });
 
     it("should parse negative sizes", function() {
-        var parse = getParsed("\\rule{-1em}{- 0.2em}")[0];
+        const parse = getParsed("\\rule{-1em}{- 0.2em}")[0];
 
         expect(parse.value.width.number).toBeCloseTo(-1);
         expect(parse.value.height.number).toBeCloseTo(-0.2);
@@ -966,17 +1000,23 @@ describe("A rule parser", function() {
 });
 
 describe("A kern parser", function() {
-    var emKern = "\\kern{1em}";
-    var exKern = "\\kern{1ex}";
-    var badUnitRule = "\\kern{1px}";
-    var noNumberRule = "\\kern{em}";
+    const emKern = "\\kern{1em}";
+    const exKern = "\\kern{1ex}";
+    const muKern = "\\kern{1mu}";
+    const abKern = "a\\kern{1em}b";
+    const badUnitRule = "\\kern{1au}";
+    const noNumberRule = "\\kern{em}";
 
     it("should list the correct units", function() {
-        var emParse = getParsed(emKern)[0];
-        var exParse = getParsed(exKern)[0];
+        const emParse = getParsed(emKern)[0];
+        const exParse = getParsed(exKern)[0];
+        const muParse = getParsed(muKern)[0];
+        const abParse = getParsed(abKern)[1];
 
         expect(emParse.value.dimension.unit).toEqual("em");
         expect(exParse.value.dimension.unit).toEqual("ex");
+        expect(muParse.value.dimension.unit).toEqual("mu");
+        expect(abParse.value.dimension.unit).toEqual("em");
     });
 
     it("should not parse invalid units", function() {
@@ -985,21 +1025,94 @@ describe("A kern parser", function() {
     });
 
     it("should parse negative sizes", function() {
-        var parse = getParsed("\\kern{-1em}")[0];
+        const parse = getParsed("\\kern{-1em}")[0];
         expect(parse.value.dimension.number).toBeCloseTo(-1);
+    });
+
+    it("should parse positive sizes", function() {
+        const parse = getParsed("\\kern{+1em}")[0];
+        expect(parse.value.dimension.number).toBeCloseTo(1);
+    });
+});
+
+describe("A non-braced kern parser", function() {
+    const emKern = "\\kern1em";
+    const exKern = "\\kern 1 ex";
+    const muKern = "\\kern 1mu";
+    const abKern1 = "a\\mkern1mub";
+    const abKern2 = "a\\kern-1mub";
+    const abKern3 = "a\\kern-1mu b";
+    const badUnitRule = "\\kern1au";
+    const noNumberRule = "\\kern em";
+
+    it("should list the correct units", function() {
+        const emParse = getParsed(emKern)[0];
+        const exParse = getParsed(exKern)[0];
+        const muParse = getParsed(muKern)[0];
+        const abParse1 = getParsed(abKern1)[1];
+        const abParse2 = getParsed(abKern2)[1];
+        const abParse3 = getParsed(abKern3)[1];
+
+        expect(emParse.value.dimension.unit).toEqual("em");
+        expect(exParse.value.dimension.unit).toEqual("ex");
+        expect(muParse.value.dimension.unit).toEqual("mu");
+        expect(abParse1.value.dimension.unit).toEqual("mu");
+        expect(abParse2.value.dimension.unit).toEqual("mu");
+        expect(abParse3.value.dimension.unit).toEqual("mu");
+    });
+
+    it("should parse elements on either side of a kern", function() {
+        const abParse1 = getParsed(abKern1);
+        const abParse2 = getParsed(abKern2);
+        const abParse3 = getParsed(abKern3);
+
+        expect(abParse1.length).toEqual(3);
+        expect(abParse1[0].value).toEqual("a");
+        expect(abParse1[2].value).toEqual("b");
+        expect(abParse2.length).toEqual(3);
+        expect(abParse2[0].value).toEqual("a");
+        expect(abParse2[2].value).toEqual("b");
+        expect(abParse3.length).toEqual(3);
+        expect(abParse3[0].value).toEqual("a");
+        expect(abParse3[2].value).toEqual("b");
+    });
+
+    it("should not parse invalid units", function() {
+        expect(badUnitRule).toNotParse();
+        expect(noNumberRule).toNotParse();
+    });
+
+    it("should parse negative sizes", function() {
+        const parse = getParsed("\\kern-1em")[0];
+        expect(parse.value.dimension.number).toBeCloseTo(-1);
+    });
+
+    it("should parse positive sizes", function() {
+        const parse = getParsed("\\kern+1em")[0];
+        expect(parse.value.dimension.number).toBeCloseTo(1);
+    });
+
+    it("should handle whitespace", function() {
+        const abKern = "a\\kern\t-\r1  \n mu\nb";
+        const abParse = getParsed(abKern);
+
+        expect(abParse.length).toEqual(3);
+        expect(abParse[0].value).toEqual("a");
+        expect(abParse[1].value.dimension.unit).toEqual("mu");
+        expect(abParse[2].value).toEqual("b");
     });
 });
 
 describe("A left/right parser", function() {
-    var normalLeftRight = "\\left( \\dfrac{x}{y} \\right)";
-    var emptyRight = "\\left( \\dfrac{x}{y} \\right.";
+    const normalLeftRight = "\\left( \\dfrac{x}{y} \\right)";
+    const emptyRight = "\\left( \\dfrac{x}{y} \\right.";
 
     it("should not fail", function() {
         expect(normalLeftRight).toParse();
     });
 
     it("should produce a leftright", function() {
-        var parse = getParsed(normalLeftRight)[0];
+        const parse = getParsed(normalLeftRight)[0];
 
         expect(parse.type).toEqual("leftright");
         expect(parse.value.left).toEqual("(");
@@ -1007,8 +1120,8 @@ describe("A left/right parser", function() {
     });
 
     it("should error when it is mismatched", function() {
-        var unmatchedLeft = "\\left( \\dfrac{x}{y}";
-        var unmatchedRight = "\\dfrac{x}{y} \\right)";
+        const unmatchedLeft = "\\left( \\dfrac{x}{y}";
+        const unmatchedRight = "\\dfrac{x}{y} \\right)";
 
         expect(unmatchedLeft).toNotParse();
 
@@ -1016,12 +1129,12 @@ describe("A left/right parser", function() {
     });
 
     it("should error when braces are mismatched", function() {
-        var unmatched = "{ \\left( \\dfrac{x}{y} } \\right)";
+        const unmatched = "{ \\left( \\dfrac{x}{y} } \\right)";
         expect(unmatched).toNotParse();
     });
 
     it("should error when non-delimiters are provided", function() {
-        var nonDelimiter = "\\left$ \\dfrac{x}{y} \\right)";
+        const nonDelimiter = "\\left$ \\dfrac{x}{y} \\right)";
         expect(nonDelimiter).toNotParse();
     });
 
@@ -1030,8 +1143,28 @@ describe("A left/right parser", function() {
     });
 
     it("should parse the '.' delimiter with normal sizes", function() {
-        var normalEmpty = "\\Bigl .";
+        const normalEmpty = "\\Bigl .";
         expect(normalEmpty).toParse();
+    });
+
+    it("should handle \\middle", function() {
+        const normalMiddle = "\\left( \\dfrac{x}{y} \\middle| \\dfrac{y}{z} \\right)";
+        expect(normalMiddle).toParse();
+    });
+
+    it("should handle multiple \\middles", function() {
+        const multiMiddle = "\\left( \\dfrac{x}{y} \\middle| \\dfrac{y}{z} \\middle/ \\dfrac{z}{q} \\right)";
+        expect(multiMiddle).toParse();
+    });
+
+    it("should handle nested \\middles", function() {
+        const nestedMiddle = "\\left( a^2 \\middle| \\left( b \\middle/ c \\right) \\right)";
+        expect(nestedMiddle).toParse();
+    });
+
+    it("should error when \\middle is not in \\left...\\right", function() {
+        const unmatchedMiddle = "(\\middle|\\dfrac{x}{y})";
+        expect(unmatchedMiddle).toNotParse();
     });
 });
 
@@ -1066,8 +1199,8 @@ describe("A begin/end parser", function() {
     });
 
     it("should nest", function() {
-        var m1 = "\\begin{pmatrix}1&2\\\\3&4\\end{pmatrix}";
-        var m2 = "\\begin{array}{rl}" + m1 + "&0\\\\0&" + m1 + "\\end{array}";
+        const m1 = "\\begin{pmatrix}1&2\\\\3&4\\end{pmatrix}";
+        const m2 = "\\begin{array}{rl}" + m1 + "&0\\\\0&" + m1 + "\\end{array}";
         expect(m2).toParse();
     });
 
@@ -1077,8 +1210,8 @@ describe("A begin/end parser", function() {
 });
 
 describe("A sqrt parser", function() {
-    var sqrt = "\\sqrt{x}";
-    var missingGroup = "\\sqrt";
+    const sqrt = "\\sqrt{x}";
+    const missingGroup = "\\sqrt";
 
     it("should parse square roots", function() {
         expect(sqrt).toParse();
@@ -1089,7 +1222,7 @@ describe("A sqrt parser", function() {
     });
 
     it("should produce sqrts", function() {
-        var parse = getParsed(sqrt)[0];
+        const parse = getParsed(sqrt)[0];
 
         expect(parse.type).toEqual("sqrt");
     });
@@ -1101,16 +1234,16 @@ describe("A TeX-compliant parser", function() {
     });
 
     it("should fail if there are not enough arguments", function() {
-        var missingGroups = [
+        const missingGroups = [
             "\\frac{x}",
-            "\\color{#fff}",
+            "\\textcolor{#fff}",
             "\\rule{1em}",
             "\\llap",
             "\\bigl",
             "\\text",
         ];
 
-        for (var i = 0; i < missingGroups.length; i++) {
+        for (let i = 0; i < missingGroups.length; i++) {
             expect(missingGroups[i]).toNotParse();
         }
     });
@@ -1121,67 +1254,67 @@ describe("A TeX-compliant parser", function() {
     });
 
     it("should fail when arguments require arguments", function() {
-        var badArguments = [
+        const badArguments = [
             "\\frac \\frac x y z",
             "\\frac x \\frac y z",
             "\\frac \\sqrt x y",
             "\\frac x \\sqrt y",
-            "\\frac \\llap x y",
-            "\\frac x \\llap y",
+            "\\frac \\mathllap x y",
+            "\\frac x \\mathllap y",
             // This actually doesn't work in real TeX, but it is suprisingly
             // hard to get this to correctly work. So, we take hit of very small
             // amounts of non-compatiblity in order for the rest of the tests to
             // work
             // "\\llap \\frac x y",
-            "\\llap \\llap x",
-            "\\sqrt \\llap x",
+            "\\mathllap \\mathllap x",
+            "\\sqrt \\mathllap x",
         ];
 
-        for (var i = 0; i < badArguments.length; i++) {
+        for (let i = 0; i < badArguments.length; i++) {
             expect(badArguments[i]).toNotParse();
         }
     });
 
     it("should work when the arguments have braces", function() {
-        var goodArguments = [
+        const goodArguments = [
             "\\frac {\\frac x y} z",
             "\\frac x {\\frac y z}",
             "\\frac {\\sqrt x} y",
             "\\frac x {\\sqrt y}",
-            "\\frac {\\llap x} y",
-            "\\frac x {\\llap y}",
-            "\\llap {\\frac x y}",
-            "\\llap {\\llap x}",
-            "\\sqrt {\\llap x}",
+            "\\frac {\\mathllap x} y",
+            "\\frac x {\\mathllap y}",
+            "\\mathllap {\\frac x y}",
+            "\\mathllap {\\mathllap x}",
+            "\\sqrt {\\mathllap x}",
         ];
 
-        for (var i = 0; i < goodArguments.length; i++) {
+        for (let i = 0; i < goodArguments.length; i++) {
             expect(goodArguments[i]).toParse();
         }
     });
 
     it("should fail when sup/subscripts require arguments", function() {
-        var badSupSubscripts = [
+        const badSupSubscripts = [
             "x^\\sqrt x",
-            "x^\\llap x",
+            "x^\\mathllap x",
             "x_\\sqrt x",
-            "x_\\llap x",
+            "x_\\mathllap x",
         ];
 
-        for (var i = 0; i < badSupSubscripts.length; i++) {
+        for (let i = 0; i < badSupSubscripts.length; i++) {
             expect(badSupSubscripts[i]).toNotParse();
         }
     });
 
     it("should work when sup/subscripts arguments have braces", function() {
-        var goodSupSubscripts = [
+        const goodSupSubscripts = [
             "x^{\\sqrt x}",
-            "x^{\\llap x}",
+            "x^{\\mathllap x}",
             "x_{\\sqrt x}",
-            "x_{\\llap x}",
+            "x_{\\mathllap x}",
         ];
 
-        for (var i = 0; i < goodSupSubscripts.length; i++) {
+        for (let i = 0; i < goodSupSubscripts.length; i++) {
             expect(goodSupSubscripts[i]).toParse();
         }
     });
@@ -1190,7 +1323,6 @@ describe("A TeX-compliant parser", function() {
         expect("x''''").toParse();
         expect("x_2''").toParse();
         expect("x''_2").toParse();
-        expect("x'_2'").toParse();
     });
 
     it("should fail when sup/subscripts are interspersed with arguments", function() {
@@ -1212,29 +1344,29 @@ describe("A TeX-compliant parser", function() {
     });
 
     it("should fail when arguments are \\left", function() {
-        var badLeftArguments = [
+        const badLeftArguments = [
             "\\frac \\left( x \\right) y",
             "\\frac x \\left( y \\right)",
-            "\\llap \\left( x \\right)",
+            "\\mathllap \\left( x \\right)",
             "\\sqrt \\left( x \\right)",
             "x^\\left( x \\right)",
         ];
 
-        for (var i = 0; i < badLeftArguments.length; i++) {
+        for (let i = 0; i < badLeftArguments.length; i++) {
             expect(badLeftArguments[i]).toNotParse();
         }
     });
 
     it("should succeed when there are braces around the \\left/\\right", function() {
-        var goodLeftArguments = [
+        const goodLeftArguments = [
             "\\frac {\\left( x \\right)} y",
             "\\frac x {\\left( y \\right)}",
-            "\\llap {\\left( x \\right)}",
+            "\\mathllap {\\left( x \\right)}",
             "\\sqrt {\\left( x \\right)}",
             "x^{\\left( x \\right)}",
         ];
 
-        for (var i = 0; i < goodLeftArguments.length; i++) {
+        for (let i = 0; i < goodLeftArguments.length; i++) {
             expect(goodLeftArguments[i]).toParse();
         }
     });
@@ -1249,22 +1381,22 @@ describe("A style change parser", function() {
     });
 
     it("should produce the correct style", function() {
-        var displayParse = getParsed("\\displaystyle x")[0];
+        const displayParse = getParsed("\\displaystyle x")[0];
         expect(displayParse.value.style).toEqual("display");
 
-        var scriptscriptParse = getParsed("\\scriptscriptstyle x")[0];
+        const scriptscriptParse = getParsed("\\scriptscriptstyle x")[0];
         expect(scriptscriptParse.value.style).toEqual("scriptscript");
     });
 
     it("should only change the style within its group", function() {
-        var text = "a b { c d \\displaystyle e f } g h";
-        var parse = getParsed(text);
+        const text = "a b { c d \\displaystyle e f } g h";
+        const parse = getParsed(text);
 
-        var displayNode = parse[2].value[2];
+        const displayNode = parse[2].value[2];
 
         expect(displayNode.type).toEqual("styling");
 
-        var displayBody = displayNode.value.value;
+        const displayBody = displayNode.value.value;
 
         expect(displayBody.length).toEqual(2);
         expect(displayBody[0].value).toEqual("e");
@@ -1287,34 +1419,34 @@ describe("A font parser", function() {
     });
 
     it("should produce the correct fonts", function() {
-        var mathbbParse = getParsed("\\mathbb x")[0];
+        const mathbbParse = getParsed("\\mathbb x")[0];
         expect(mathbbParse.value.font).toEqual("mathbb");
         expect(mathbbParse.value.type).toEqual("font");
 
-        var mathrmParse = getParsed("\\mathrm x")[0];
+        const mathrmParse = getParsed("\\mathrm x")[0];
         expect(mathrmParse.value.font).toEqual("mathrm");
         expect(mathrmParse.value.type).toEqual("font");
 
-        var mathitParse = getParsed("\\mathit x")[0];
+        const mathitParse = getParsed("\\mathit x")[0];
         expect(mathitParse.value.font).toEqual("mathit");
         expect(mathitParse.value.type).toEqual("font");
 
-        var mathcalParse = getParsed("\\mathcal C")[0];
+        const mathcalParse = getParsed("\\mathcal C")[0];
         expect(mathcalParse.value.font).toEqual("mathcal");
         expect(mathcalParse.value.type).toEqual("font");
 
-        var mathfrakParse = getParsed("\\mathfrak C")[0];
+        const mathfrakParse = getParsed("\\mathfrak C")[0];
         expect(mathfrakParse.value.font).toEqual("mathfrak");
         expect(mathfrakParse.value.type).toEqual("font");
     });
 
     it("should parse nested font commands", function() {
-        var nestedParse = getParsed("\\mathbb{R \\neq \\mathrm{R}}")[0];
+        const nestedParse = getParsed("\\mathbb{R \\neq \\mathrm{R}}")[0];
         expect(nestedParse.value.font).toEqual("mathbb");
         expect(nestedParse.value.type).toEqual("font");
 
         expect(nestedParse.value.body.value.length).toEqual(3);
-        var bbBody = nestedParse.value.body.value;
+        const bbBody = nestedParse.value.body.value;
         expect(bbBody[0].type).toEqual("mathord");
         expect(bbBody[1].type).toEqual("rel");
         expect(bbBody[2].type).toEqual("font");
@@ -1322,11 +1454,11 @@ describe("A font parser", function() {
         expect(bbBody[2].value.type).toEqual("font");
     });
 
-    it("should work with \\color", function() {
-        var colorMathbbParse = getParsed("\\color{blue}{\\mathbb R}")[0];
+    it("should work with \\textcolor", function() {
+        const colorMathbbParse = getParsed("\\textcolor{blue}{\\mathbb R}")[0];
         expect(colorMathbbParse.value.type).toEqual("color");
         expect(colorMathbbParse.value.color).toEqual("blue");
-        var body = colorMathbbParse.value.value;
+        const body = colorMathbbParse.value.value;
         expect(body.length).toEqual(1);
         expect(body[0].value.type).toEqual("font");
         expect(body[0].value.font).toEqual("mathbb");
@@ -1337,7 +1469,7 @@ describe("A font parser", function() {
     });
 
     it("should nest fonts correctly", function() {
-        var bf = getParsed("\\mathbf{a\\mathrm{b}c}")[0];
+        const bf = getParsed("\\mathbf{a\\mathrm{b}c}")[0];
         expect(bf.value.type).toEqual("font");
         expect(bf.value.font).toEqual("mathbf");
         expect(bf.value.body.value.length).toEqual(3);
@@ -1354,43 +1486,107 @@ describe("A font parser", function() {
 
 describe("An HTML font tree-builder", function() {
     it("should render \\mathbb{R} with the correct font", function() {
-        var markup = katex.renderToString("\\mathbb{R}");
+        const markup = katex.renderToString("\\mathbb{R}");
         expect(markup).toContain("<span class=\"mord mathbb\">R</span>");
     });
 
     it("should render \\mathrm{R} with the correct font", function() {
-        var markup = katex.renderToString("\\mathrm{R}");
+        const markup = katex.renderToString("\\mathrm{R}");
         expect(markup).toContain("<span class=\"mord mathrm\">R</span>");
     });
 
     it("should render \\mathcal{R} with the correct font", function() {
-        var markup = katex.renderToString("\\mathcal{R}");
+        const markup = katex.renderToString("\\mathcal{R}");
         expect(markup).toContain("<span class=\"mord mathcal\">R</span>");
     });
 
     it("should render \\mathfrak{R} with the correct font", function() {
-        var markup = katex.renderToString("\\mathfrak{R}");
+        const markup = katex.renderToString("\\mathfrak{R}");
         expect(markup).toContain("<span class=\"mord mathfrak\">R</span>");
     });
 
+    it("should render \\text{R} with the correct font", function() {
+        const markup = katex.renderToString("\\text{R}");
+        expect(markup).toContain("<span class=\"mord mathrm\">R</span>");
+    });
+
+    it("should render \\textit{R} with the correct font", function() {
+        const markup = katex.renderToString("\\textit{R}");
+        expect(markup).toContain("<span class=\"mord textit\">R</span>");
+    });
+
+    it("should render \\text{\\textit{R}} with the correct font", function() {
+        const markup = katex.renderToString("\\text{\\textit{R}}");
+        expect(markup).toContain("<span class=\"mord textit\">R</span>");
+    });
+
+    it("should render \\text{R\\textit{S}T} with the correct fonts", function() {
+        const markup = katex.renderToString("\\text{R\\textit{S}T}");
+        expect(markup).toContain("<span class=\"mord mathrm\">R</span>");
+        expect(markup).toContain("<span class=\"mord textit\">S</span>");
+        expect(markup).toContain("<span class=\"mord mathrm\">T</span>");
+    });
+
+    it("should render \\textbf{R} with the correct font", function() {
+        const markup = katex.renderToString("\\textbf{R}");
+        expect(markup).toContain("<span class=\"mord mathbf\">R</span>");
+    });
+
+    it("should render \\textsf{R} with the correct font", function() {
+        const markup = katex.renderToString("\\textsf{R}");
+        expect(markup).toContain("<span class=\"mord mathsf\">R</span>");
+    });
+
+    it("should render \\texttt{R} with the correct font", function() {
+        const markup = katex.renderToString("\\texttt{R}");
+        expect(markup).toContain("<span class=\"mord mathtt\">R</span>");
+    });
+
     it("should render a combination of font and color changes", function() {
-        var markup = katex.renderToString("\\color{blue}{\\mathbb R}");
-        var span = "<span class=\"mord mathbb\" style=\"color:blue;\">R</span>";
+        let markup = katex.renderToString("\\textcolor{blue}{\\mathbb R}");
+        let span = "<span class=\"mord mathbb\" style=\"color:blue;\">R</span>";
         expect(markup).toContain(span);
 
-        markup = katex.renderToString("\\mathbb{\\color{blue}{R}}");
+        markup = katex.renderToString("\\mathbb{\\textcolor{blue}{R}}");
         span = "<span class=\"mord mathbb\" style=\"color:blue;\">R</span>";
         expect(markup).toContain(span);
+    });
+
+    it("should throw TypeError when the expression is of the wrong type", function() {
+        expect(function() {
+            katex.renderToString({badInputType: "yes"});
+        }).toThrowError(TypeError);
+        expect(function() {
+            katex.renderToString([1, 2]);
+        }).toThrowError(TypeError);
+        expect(function() {
+            katex.renderToString(undefined);
+        }).toThrowError(TypeError);
+        expect(function() {
+            katex.renderToString(null);
+        }).toThrowError(TypeError);
+        expect(function() {
+            katex.renderToString(1.234);
+        }).toThrowError(TypeError);
+    });
+
+    it("should not throw TypeError when the expression is a supported type", function() {
+        expect(function() {
+            katex.renderToString("\\sqrt{123}");
+        }).not.toThrowError(TypeError);
+        expect(function() {
+            katex.renderToString(new String("\\sqrt{123}"));
+        }).not.toThrowError(TypeError);
     });
 });
 
 
 describe("A MathML font tree-builder", function() {
-    var contents = "Ax2k\\omega\\Omega\\imath+";
+    const contents = "Ax2k\\omega\\Omega\\imath+";
 
     it("should render " + contents + " with the correct mathvariants", function() {
-        var tree = getParsed(contents);
-        var markup = buildMathML(tree, contents, defaultOptions).toMarkup();
+        const tree = getParsed(contents);
+        const markup = buildMathML(tree, contents, defaultOptions).toMarkup();
         expect(markup).toContain("<mi>A</mi>");
         expect(markup).toContain("<mi>x</mi>");
         expect(markup).toContain("<mn>2</mn>");
@@ -1401,12 +1597,12 @@ describe("A MathML font tree-builder", function() {
     });
 
     it("should render \\mathbb{" + contents + "} with the correct mathvariants", function() {
-        var tex = "\\mathbb{" + contents + "}";
-        var tree = getParsed(tex);
-        var markup = buildMathML(tree, tex, defaultOptions).toMarkup();
+        const tex = "\\mathbb{" + contents + "}";
+        const tree = getParsed(tex);
+        const markup = buildMathML(tree, tex, defaultOptions).toMarkup();
         expect(markup).toContain("<mi mathvariant=\"double-struck\">A</mi>");
         expect(markup).toContain("<mi>x</mi>");
-        expect(markup).toContain("<mn mathvariant=\"normal\">2</mn>");
+        expect(markup).toContain("<mn>2</mn>");
         expect(markup).toContain("<mi>\u03c9</mi>");                        // \omega
         expect(markup).toContain("<mi mathvariant=\"normal\">\u03A9</mi>"); // \Omega
         expect(markup).toContain("<mi>\u0131</mi>");                        // \imath
@@ -1414,12 +1610,12 @@ describe("A MathML font tree-builder", function() {
     });
 
     it("should render \\mathrm{" + contents + "} with the correct mathvariants", function() {
-        var tex = "\\mathrm{" + contents + "}";
-        var tree = getParsed(tex);
-        var markup = buildMathML(tree, tex, defaultOptions).toMarkup();
+        const tex = "\\mathrm{" + contents + "}";
+        const tree = getParsed(tex);
+        const markup = buildMathML(tree, tex, defaultOptions).toMarkup();
         expect(markup).toContain("<mi mathvariant=\"normal\">A</mi>");
         expect(markup).toContain("<mi mathvariant=\"normal\">x</mi>");
-        expect(markup).toContain("<mn mathvariant=\"normal\">2</mn>");
+        expect(markup).toContain("<mn>2</mn>");
         expect(markup).toContain("<mi>\u03c9</mi>");   // \omega
         expect(markup).toContain("<mi mathvariant=\"normal\">\u03A9</mi>");   // \Omega
         expect(markup).toContain("<mi>\u0131</mi>");   // \imath
@@ -1427,22 +1623,22 @@ describe("A MathML font tree-builder", function() {
     });
 
     it("should render \\mathit{" + contents + "} with the correct mathvariants", function() {
-        var tex = "\\mathit{" + contents + "}";
-        var tree = getParsed(tex);
-        var markup = buildMathML(tree, tex, defaultOptions).toMarkup();
-        expect(markup).toContain("<mi mathvariant=\"italic\">A</mi>");
-        expect(markup).toContain("<mi mathvariant=\"italic\">x</mi>");
+        const tex = "\\mathit{" + contents + "}";
+        const tree = getParsed(tex);
+        const markup = buildMathML(tree, tex, defaultOptions).toMarkup();
+        expect(markup).toContain("<mi>A</mi>");
+        expect(markup).toContain("<mi>x</mi>");
         expect(markup).toContain("<mn mathvariant=\"italic\">2</mn>");
-        expect(markup).toContain("<mi mathvariant=\"italic\">\u03c9</mi>");   // \omega
-        expect(markup).toContain("<mi mathvariant=\"italic\">\u03A9</mi>");   // \Omega
-        expect(markup).toContain("<mi mathvariant=\"italic\">\u0131</mi>");   // \imath
+        expect(markup).toContain("<mi>\u03c9</mi>");   // \omega
+        expect(markup).toContain("<mi>\u03A9</mi>");   // \Omega
+        expect(markup).toContain("<mi>\u0131</mi>");   // \imath
         expect(markup).toContain("<mo>+</mo>");
     });
 
     it("should render \\mathbf{" + contents + "} with the correct mathvariants", function() {
-        var tex = "\\mathbf{" + contents + "}";
-        var tree = getParsed(tex);
-        var markup = buildMathML(tree, tex, defaultOptions).toMarkup();
+        const tex = "\\mathbf{" + contents + "}";
+        const tree = getParsed(tex);
+        const markup = buildMathML(tree, tex, defaultOptions).toMarkup();
         expect(markup).toContain("<mi mathvariant=\"bold\">A</mi>");
         expect(markup).toContain("<mi mathvariant=\"bold\">x</mi>");
         expect(markup).toContain("<mn mathvariant=\"bold\">2</mn>");
@@ -1453,9 +1649,9 @@ describe("A MathML font tree-builder", function() {
     });
 
     it("should render \\mathcal{" + contents + "} with the correct mathvariants", function() {
-        var tex = "\\mathcal{" + contents + "}";
-        var tree = getParsed(tex);
-        var markup = buildMathML(tree, tex, defaultOptions).toMarkup();
+        const tex = "\\mathcal{" + contents + "}";
+        const tree = getParsed(tex);
+        const markup = buildMathML(tree, tex, defaultOptions).toMarkup();
         expect(markup).toContain("<mi mathvariant=\"script\">A</mi>");
         expect(markup).toContain("<mi>x</mi>");                             // script is caps only
         expect(markup).toContain("<mn mathvariant=\"script\">2</mn>");
@@ -1468,9 +1664,9 @@ describe("A MathML font tree-builder", function() {
     });
 
     it("should render \\mathfrak{" + contents + "} with the correct mathvariants", function() {
-        var tex = "\\mathfrak{" + contents + "}";
-        var tree = getParsed(tex);
-        var markup = buildMathML(tree, tex, defaultOptions).toMarkup();
+        const tex = "\\mathfrak{" + contents + "}";
+        const tree = getParsed(tex);
+        const markup = buildMathML(tree, tex, defaultOptions).toMarkup();
         expect(markup).toContain("<mi mathvariant=\"fraktur\">A</mi>");
         expect(markup).toContain("<mi mathvariant=\"fraktur\">x</mi>");
         expect(markup).toContain("<mn mathvariant=\"fraktur\">2</mn>");
@@ -1483,14 +1679,14 @@ describe("A MathML font tree-builder", function() {
     });
 
     it("should render \\mathscr{" + contents + "} with the correct mathvariants", function() {
-        var tex = "\\mathscr{" + contents + "}";
-        var tree = getParsed(tex);
-        var markup = buildMathML(tree, tex, defaultOptions).toMarkup();
+        const tex = "\\mathscr{" + contents + "}";
+        const tree = getParsed(tex);
+        const markup = buildMathML(tree, tex, defaultOptions).toMarkup();
         expect(markup).toContain("<mi mathvariant=\"script\">A</mi>");
         // MathJax marks everything below as "script" except \omega
         // We don't have these glyphs in "script" and neither does MathJax
         expect(markup).toContain("<mi>x</mi>");
-        expect(markup).toContain("<mn mathvariant=\"normal\">2</mn>");
+        expect(markup).toContain("<mn>2</mn>");
         expect(markup).toContain("<mi>\u03c9</mi>");                        // \omega
         expect(markup).toContain("<mi mathvariant=\"normal\">\u03A9</mi>"); // \Omega
         expect(markup).toContain("<mi>\u0131</mi>");                        // \imath
@@ -1498,9 +1694,9 @@ describe("A MathML font tree-builder", function() {
     });
 
     it("should render \\mathsf{" + contents + "} with the correct mathvariants", function() {
-        var tex = "\\mathsf{" + contents + "}";
-        var tree = getParsed(tex);
-        var markup = buildMathML(tree, tex, defaultOptions).toMarkup();
+        const tex = "\\mathsf{" + contents + "}";
+        const tree = getParsed(tex);
+        const markup = buildMathML(tree, tex, defaultOptions).toMarkup();
         expect(markup).toContain("<mi mathvariant=\"sans-serif\">A</mi>");
         expect(markup).toContain("<mi mathvariant=\"sans-serif\">x</mi>");
         expect(markup).toContain("<mn mathvariant=\"sans-serif\">2</mn>");
@@ -1511,16 +1707,16 @@ describe("A MathML font tree-builder", function() {
     });
 
     it("should render a combination of font and color changes", function() {
-        var tex = "\\color{blue}{\\mathbb R}";
-        var tree = getParsed(tex);
-        var markup = buildMathML(tree, tex, defaultOptions).toMarkup();
-        var node = "<mstyle mathcolor=\"blue\">" +
+        let tex = "\\textcolor{blue}{\\mathbb R}";
+        let tree = getParsed(tex);
+        let markup = buildMathML(tree, tex, defaultOptions).toMarkup();
+        let node = "<mstyle mathcolor=\"blue\">" +
             "<mi mathvariant=\"double-struck\">R</mi>" +
             "</mstyle>";
         expect(markup).toContain(node);
 
         // reverse the order of the commands
-        tex = "\\mathbb{\\color{blue}{R}}";
+        tex = "\\mathbb{\\textcolor{blue}{R}}";
         tree = getParsed(tex);
         markup = buildMathML(tree, tex, defaultOptions).toMarkup();
         node = "<mstyle mathcolor=\"blue\">" +
@@ -1528,17 +1724,33 @@ describe("A MathML font tree-builder", function() {
             "</mstyle>";
         expect(markup).toContain(node);
     });
+
+    it("should render text as <mtext>", function() {
+        const tex = "\\text{for }";
+        const tree = getParsed(tex);
+        const markup = buildMathML(tree, tex, defaultOptions).toMarkup();
+        expect(markup).toContain("<mtext>for\u00a0</mtext>");
+    });
+
+    it("should render math within text as side-by-side children", function() {
+        const tex = "\\text{graph: $y = mx + b$}";
+        const tree = getParsed(tex);
+        const markup = buildMathML(tree, tex, defaultOptions).toMarkup();
+        expect(markup).toContain("<mrow><mtext>graph:\u00a0</mtext>");
+        expect(markup).toContain(
+            "<mi>y</mi><mo>=</mo><mi>m</mi><mi>x</mi><mo>+</mo><mi>b</mi>");
+    });
 });
 
 describe("A bin builder", function() {
     it("should create mbins normally", function() {
-        var built = getBuilt("x + y");
+        const built = getBuilt("x + y");
 
         expect(built[1].classes).toContain("mbin");
     });
 
     it("should create ords when at the beginning of lists", function() {
-        var built = getBuilt("+ x");
+        const built = getBuilt("+ x");
 
         expect(built[0].classes).toContain("mord");
         expect(built[0].classes).not.toContain("mbin");
@@ -1554,14 +1766,15 @@ describe("A bin builder", function() {
 
     it("should correctly interact with color objects", function() {
         expect(getBuilt("\\blue{x}+y")[1].classes).toContain("mbin");
-        expect(getBuilt("\\blue{x+}+y")[1].classes).toContain("mord");
+        expect(getBuilt("\\blue{x+}+y")[1].classes).toContain("mbin");
+        expect(getBuilt("\\blue{x+}+y")[2].classes).toContain("mord");
     });
 });
 
 describe("A markup generator", function() {
     it("marks trees up", function() {
         // Just a few quick sanity checks here...
-        var markup = katex.renderToString("\\sigma^2");
+        const markup = katex.renderToString("\\sigma^2");
         expect(markup.indexOf("<span")).toBe(0);
         expect(markup).toContain("\u03c3");  // sigma
         expect(markup).toContain("margin-right");
@@ -1569,7 +1782,7 @@ describe("A markup generator", function() {
     });
 
     it("generates both MathML and HTML", function() {
-        var markup = katex.renderToString("a");
+        const markup = katex.renderToString("a");
 
         expect(markup).toContain("<span");
         expect(markup).toContain("<math");
@@ -1578,7 +1791,7 @@ describe("A markup generator", function() {
 
 describe("A parse tree generator", function() {
     it("generates a tree", function() {
-        var tree = stripPositions(katex.__parse("\\sigma^2"));
+        const tree = stripPositions(katex.__parse("\\sigma^2"));
         expect(JSON.stringify(tree)).toEqual(JSON.stringify([
             {
                 "type": "supsub",
@@ -1610,19 +1823,23 @@ describe("An accent parser", function() {
     });
 
     it("should produce accents", function() {
-        var parse = getParsed("\\vec x")[0];
+        const parse = getParsed("\\vec x")[0];
 
         expect(parse.type).toEqual("accent");
     });
 
     it("should be grouped more tightly than supsubs", function() {
-        var parse = getParsed("\\vec x^2")[0];
+        const parse = getParsed("\\vec x^2")[0];
 
         expect(parse.type).toEqual("supsub");
     });
 
-    it("should not parse expanding accents", function() {
-        expect("\\widehat{x}").toNotParse();
+    it("should parse stretchy, shifty accents", function() {
+        expect("\\widehat{x}").toParse();
+    });
+
+    it("should parse stretchy, non-shifty accents", function() {
+        expect("\\overrightarrow{x}").toParse();
     });
 });
 
@@ -1643,16 +1860,320 @@ describe("An accent builder", function() {
     });
 });
 
+describe("A stretchy and shifty accent builder", function() {
+    it("should not fail", function() {
+        expect("\\widehat{AB}").toBuild();
+        expect("\\widehat{AB}^2").toBuild();
+        expect("\\widehat{AB}_2").toBuild();
+        expect("\\widehat{AB}_2^2").toBuild();
+    });
+
+    it("should produce mords", function() {
+        expect(getBuilt("\\widehat{AB}")[0].classes).toContain("mord");
+        expect(getBuilt("\\widehat +")[0].classes).toContain("mord");
+        expect(getBuilt("\\widehat +")[0].classes).not.toContain("mbin");
+        expect(getBuilt("\\widehat )^2")[0].classes).toContain("mord");
+        expect(getBuilt("\\widehat )^2")[0].classes).not.toContain("mclose");
+    });
+});
+
+describe("A stretchy and non-shifty accent builder", function() {
+    it("should not fail", function() {
+        expect("\\overrightarrow{AB}").toBuild();
+        expect("\\overrightarrow{AB}^2").toBuild();
+        expect("\\overrightarrow{AB}_2").toBuild();
+        expect("\\overrightarrow{AB}_2^2").toBuild();
+    });
+
+    it("should produce mords", function() {
+        expect(getBuilt("\\overrightarrow{AB}")[0].classes).toContain("mord");
+        expect(getBuilt("\\overrightarrow +")[0].classes).toContain("mord");
+        expect(getBuilt("\\overrightarrow +")[0].classes).not.toContain("mbin");
+        expect(getBuilt("\\overrightarrow )^2")[0].classes).toContain("mord");
+        expect(getBuilt("\\overrightarrow )^2")[0].classes).not.toContain("mclose");
+    });
+});
+
+describe("An under-accent parser", function() {
+    it("should not fail", function() {
+        expect("\\underrightarrow{x}").toParse();
+        expect("\\underrightarrow{x^2}").toParse();
+        expect("\\underrightarrow{x}^2").toParse();
+        expect("\\underrightarrow x").toParse();
+    });
+
+    it("should produce accentUnder", function() {
+        const parse = getParsed("\\underrightarrow x")[0];
+
+        expect(parse.type).toEqual("accentUnder");
+    });
+
+    it("should be grouped more tightly than supsubs", function() {
+        const parse = getParsed("\\underrightarrow x^2")[0];
+
+        expect(parse.type).toEqual("supsub");
+    });
+});
+
+describe("An under-accent builder", function() {
+    it("should not fail", function() {
+        expect("\\underrightarrow{x}").toBuild();
+        expect("\\underrightarrow{x}^2").toBuild();
+        expect("\\underrightarrow{x}_2").toBuild();
+        expect("\\underrightarrow{x}_2^2").toBuild();
+    });
+
+    it("should produce mords", function() {
+        expect(getBuilt("\\underrightarrow x")[0].classes).toContain("mord");
+        expect(getBuilt("\\underrightarrow +")[0].classes).toContain("mord");
+        expect(getBuilt("\\underrightarrow +")[0].classes).not.toContain("mbin");
+        expect(getBuilt("\\underrightarrow )^2")[0].classes).toContain("mord");
+        expect(getBuilt("\\underrightarrow )^2")[0].classes).not.toContain("mclose");
+    });
+});
+
+describe("An extensible arrow parser", function() {
+    it("should not fail", function() {
+        expect("\\xrightarrow{x}").toParse();
+        expect("\\xrightarrow{x^2}").toParse();
+        expect("\\xrightarrow{x}^2").toParse();
+        expect("\\xrightarrow x").toParse();
+        expect("\\xrightarrow[under]{over}").toParse();
+    });
+
+    it("should produce xArrow", function() {
+        const parse = getParsed("\\xrightarrow x")[0];
+
+        expect(parse.type).toEqual("xArrow");
+    });
+
+    it("should be grouped more tightly than supsubs", function() {
+        const parse = getParsed("\\xrightarrow x^2")[0];
+
+        expect(parse.type).toEqual("supsub");
+    });
+});
+
+describe("An extensible arrow builder", function() {
+    it("should not fail", function() {
+        expect("\\xrightarrow{x}").toBuild();
+        expect("\\xrightarrow{x}^2").toBuild();
+        expect("\\xrightarrow{x}_2").toBuild();
+        expect("\\xrightarrow{x}_2^2").toBuild();
+        expect("\\xrightarrow[under]{over}").toBuild();
+    });
+
+    it("should produce mrell", function() {
+        expect(getBuilt("\\xrightarrow x")[0].classes).toContain("mrel");
+        expect(getBuilt("\\xrightarrow [under]{over}")[0].classes).toContain("mrel");
+        expect(getBuilt("\\xrightarrow +")[0].classes).toContain("mrel");
+        expect(getBuilt("\\xrightarrow +")[0].classes).not.toContain("mbin");
+        expect(getBuilt("\\xrightarrow )^2")[0].classes).toContain("mrel");
+        expect(getBuilt("\\xrightarrow )^2")[0].classes).not.toContain("mclose");
+    });
+});
+
+describe("A horizontal brace parser", function() {
+    it("should not fail", function() {
+        expect("\\overbrace{x}").toParse();
+        expect("\\overbrace{x^2}").toParse();
+        expect("\\overbrace{x}^2").toParse();
+        expect("\\overbrace x").toParse();
+        expect("\\underbrace{x}_2").toParse();
+        expect("\\underbrace{x}_2^2").toParse();
+    });
+
+    it("should produce horizBrace", function() {
+        const parse = getParsed("\\overbrace x")[0];
+
+        expect(parse.type).toEqual("horizBrace");
+    });
+
+    it("should be grouped more tightly than supsubs", function() {
+        const parse = getParsed("\\overbrace x^2")[0];
+
+        expect(parse.type).toEqual("supsub");
+    });
+});
+
+describe("A horizontal brace builder", function() {
+    it("should not fail", function() {
+        expect("\\overbrace{x}").toBuild();
+        expect("\\overbrace{x}^2").toBuild();
+        expect("\\underbrace{x}_2").toBuild();
+        expect("\\underbrace{x}_2^2").toBuild();
+    });
+
+    it("should produce mords", function() {
+        expect(getBuilt("\\overbrace x")[0].classes).toContain("mord");
+        expect(getBuilt("\\overbrace{x}^2")[0].classes).toContain("mord");
+        expect(getBuilt("\\overbrace +")[0].classes).toContain("mord");
+        expect(getBuilt("\\overbrace +")[0].classes).not.toContain("mbin");
+        expect(getBuilt("\\overbrace )^2")[0].classes).toContain("mord");
+        expect(getBuilt("\\overbrace )^2")[0].classes).not.toContain("mclose");
+    });
+});
+
+describe("A boxed parser", function() {
+    it("should not fail", function() {
+        expect("\\boxed{x}").toParse();
+        expect("\\boxed{x^2}").toParse();
+        expect("\\boxed{x}^2").toParse();
+        expect("\\boxed x").toParse();
+    });
+
+    it("should produce enclose", function() {
+        const parse = getParsed("\\boxed x")[0];
+
+        expect(parse.type).toEqual("enclose");
+    });
+});
+
+describe("A boxed builder", function() {
+    it("should not fail", function() {
+        expect("\\boxed{x}").toBuild();
+        expect("\\boxed{x}^2").toBuild();
+        expect("\\boxed{x}_2").toBuild();
+        expect("\\boxed{x}_2^2").toBuild();
+    });
+
+    it("should produce mords", function() {
+        expect(getBuilt("\\boxed x")[0].classes).toContain("mord");
+        expect(getBuilt("\\boxed +")[0].classes).toContain("mord");
+        expect(getBuilt("\\boxed +")[0].classes).not.toContain("mbin");
+        expect(getBuilt("\\boxed )^2")[0].classes).toContain("mord");
+        expect(getBuilt("\\boxed )^2")[0].classes).not.toContain("mclose");
+    });
+});
+
+describe("A colorbox parser", function() {
+    it("should not fail, given a text argument", function() {
+        expect("\\colorbox{red}{a b}").toParse();
+        expect("\\colorbox{red}{x}^2").toParse();
+        expect("\\colorbox{red} x").toParse();
+    });
+
+    it("should fail, given a math argument", function() {
+        expect("\\colorbox{red}{\\alpha}").toNotParse();
+        expect("\\colorbox{red}{\\frac{a}{b}}").toNotParse();
+    });
+
+    it("should parse a color", function() {
+        expect("\\colorbox{red}{a b}").toParse();
+        expect("\\colorbox{#197}{a b}").toParse();
+        expect("\\colorbox{#1a9b7c}{a b}").toParse();
+    });
+
+    it("should produce enclose", function() {
+        const parse = getParsed("\\colorbox{red} x")[0];
+        expect(parse.type).toEqual("enclose");
+    });
+});
+
+describe("A colorbox builder", function() {
+    it("should not fail", function() {
+        expect("\\colorbox{red}{a b}").toBuild();
+        expect("\\colorbox{red}{a b}^2").toBuild();
+        expect("\\colorbox{red} x").toBuild();
+    });
+
+    it("should produce mords", function() {
+        expect(getBuilt("\\colorbox{red}{a b}")[0].classes).toContain("mord");
+    });
+});
+
+describe("An fcolorbox parser", function() {
+    it("should not fail, given a text argument", function() {
+        expect("\\fcolorbox{blue}{yellow}{a b}").toParse();
+        expect("\\fcolorbox{blue}{yellow}{x}^2").toParse();
+        expect("\\fcolorbox{blue}{yellow} x").toParse();
+    });
+
+    it("should fail, given a math argument", function() {
+        expect("\\fcolorbox{blue}{yellow}{\\alpha}").toNotParse();
+        expect("\\fcolorbox{blue}{yellow}{\\frac{a}{b}}").toNotParse();
+    });
+
+    it("should parse a color", function() {
+        expect("\\fcolorbox{blue}{yellow}{a b}").toParse();
+        expect("\\fcolorbox{blue}{#197}{a b}").toParse();
+        expect("\\fcolorbox{blue}{#1a9b7c}{a b}").toParse();
+    });
+
+    it("should produce enclose", function() {
+        const parse = getParsed("\\fcolorbox{blue}{yellow} x")[0];
+        expect(parse.type).toEqual("enclose");
+    });
+});
+
+describe("A fcolorbox builder", function() {
+    it("should not fail", function() {
+        expect("\\fcolorbox{blue}{yellow}{a b}").toBuild();
+        expect("\\fcolorbox{blue}{yellow}{a b}^2").toBuild();
+        expect("\\fcolorbox{blue}{yellow} x").toBuild();
+    });
+
+    it("should produce mords", function() {
+        expect(getBuilt("\\colorbox{red}{a b}")[0].classes).toContain("mord");
+    });
+});
+
+describe("A strike-through parser", function() {
+    it("should not fail", function() {
+        expect("\\cancel{x}").toParse();
+        expect("\\cancel{x^2}").toParse();
+        expect("\\cancel{x}^2").toParse();
+        expect("\\cancel x").toParse();
+    });
+
+    it("should produce enclose", function() {
+        const parse = getParsed("\\cancel x")[0];
+
+        expect(parse.type).toEqual("enclose");
+    });
+
+    it("should be grouped more tightly than supsubs", function() {
+        const parse = getParsed("\\cancel x^2")[0];
+
+        expect(parse.type).toEqual("supsub");
+    });
+});
+
+describe("A strike-through builder", function() {
+    it("should not fail", function() {
+        expect("\\cancel{x}").toBuild();
+        expect("\\cancel{x}^2").toBuild();
+        expect("\\cancel{x}_2").toBuild();
+        expect("\\cancel{x}_2^2").toBuild();
+        expect("\\sout{x}").toBuild();
+        expect("\\sout{x}^2").toBuild();
+        expect("\\sout{x}_2").toBuild();
+        expect("\\sout{x}_2^2").toBuild();
+    });
+
+    it("should produce mords", function() {
+        expect(getBuilt("\\cancel x")[0].classes).toContain("mord");
+        expect(getBuilt("\\cancel +")[0].classes).toContain("mord");
+        expect(getBuilt("\\cancel +")[0].classes).not.toContain("mbin");
+        expect(getBuilt("\\cancel )^2")[0].classes).toContain("mord");
+        expect(getBuilt("\\cancel )^2")[0].classes).not.toContain("mclose");
+    });
+});
+
 describe("A phantom parser", function() {
     it("should not fail", function() {
         expect("\\phantom{x}").toParse();
         expect("\\phantom{x^2}").toParse();
         expect("\\phantom{x}^2").toParse();
         expect("\\phantom x").toParse();
+        expect("\\hphantom{x}").toParse();
+        expect("\\hphantom{x^2}").toParse();
+        expect("\\hphantom{x}^2").toParse();
+        expect("\\hphantom x").toParse();
     });
 
     it("should build a phantom node", function() {
-        var parse = getParsed("\\phantom{x}")[0];
+        const parse = getParsed("\\phantom{x}")[0];
 
         expect(parse.type).toEqual("phantom");
         expect(parse.value.value).toBeDefined();
@@ -1665,20 +2186,64 @@ describe("A phantom builder", function() {
         expect("\\phantom{x^2}").toBuild();
         expect("\\phantom{x}^2").toBuild();
         expect("\\phantom x").toBuild();
+
+        expect("\\hphantom{x}").toBuild();
+        expect("\\hphantom{x^2}").toBuild();
+        expect("\\hphantom{x}^2").toBuild();
+        expect("\\hphantom x").toBuild();
     });
 
     it("should make the children transparent", function() {
-        var children = getBuilt("\\phantom{x+1}")[0].children;
+        const children = getBuilt("\\phantom{x+1}");
         expect(children[0].style.color).toBe("transparent");
         expect(children[1].style.color).toBe("transparent");
         expect(children[2].style.color).toBe("transparent");
     });
 
     it("should make all descendants transparent", function() {
-        var children = getBuilt("\\phantom{x+\\blue{1}}")[0].children;
+        const children = getBuilt("\\phantom{x+\\blue{1}}");
         expect(children[0].style.color).toBe("transparent");
         expect(children[1].style.color).toBe("transparent");
-        expect(children[2].children[0].style.color).toBe("transparent");
+        expect(children[2].style.color).toBe("transparent");
+    });
+});
+
+describe("A smash parser", function() {
+    it("should not fail", function() {
+        expect("\\smash{x}").toParse();
+        expect("\\smash{x^2}").toParse();
+        expect("\\smash{x}^2").toParse();
+        expect("\\smash x").toParse();
+
+        expect("\\smash[b]{x}").toParse();
+        expect("\\smash[b]{x^2}").toParse();
+        expect("\\smash[b]{x}^2").toParse();
+        expect("\\smash[b] x").toParse();
+
+        expect("\\smash[]{x}").toParse();
+        expect("\\smash[]{x^2}").toParse();
+        expect("\\smash[]{x}^2").toParse();
+        expect("\\smash[] x").toParse();
+    });
+
+    it("should build a smash node", function() {
+        const parse = getParsed("\\smash{x}")[0];
+
+        expect(parse.type).toEqual("smash");
+    });
+});
+
+describe("A smash builder", function() {
+    it("should not fail", function() {
+        expect("\\smash{x}").toBuild();
+        expect("\\smash{x^2}").toBuild();
+        expect("\\smash{x}^2").toBuild();
+        expect("\\smash x").toBuild();
+
+        expect("\\smash[b]{x}").toBuild();
+        expect("\\smash[b]{x^2}").toBuild();
+        expect("\\smash[b]{x}^2").toBuild();
+        expect("\\smash[b] x").toBuild();
     });
 });
 
@@ -1722,7 +2287,7 @@ describe("An optional argument parser", function() {
 describe("An array environment", function() {
 
     it("should accept a single alignment character", function() {
-        var parse = getParsed("\\begin{array}r1\\\\20\\end{array}");
+        const parse = getParsed("\\begin{array}r1\\\\20\\end{array}");
         expect(parse[0].type).toBe("array");
         expect(parse[0].value.cols).toEqual([
             { type: "align", align: "r" },
@@ -1730,7 +2295,7 @@ describe("An array environment", function() {
     });
 
     it("should accept vertical separators", function() {
-        var parse = getParsed("\\begin{array}{|l||c|}\\end{array}");
+        const parse = getParsed("\\begin{array}{|l||c|}\\end{array}");
         expect(parse[0].type).toBe("array");
         expect(parse[0].value.cols).toEqual([
             { type: "separator", separator: "|" },
@@ -1763,55 +2328,11 @@ describe("An aligned environment", function() {
 
 });
 
-var getMathML = function(expr, settings) {
-    var usedSettings = settings ? settings : defaultSettings;
-
-    expect(expr).toParse(usedSettings);
-
-    var built = buildMathML(parseTree(expr, usedSettings), expr, usedSettings);
-
-    // Strip off the surrounding <span>
-    return built.children[0];
-};
-
-describe("A MathML builder", function() {
-    it("should generate math nodes", function() {
-        var node = getMathML("x^2");
-
-        expect(node.type).toEqual("math");
-    });
-
-    it("should generate appropriate MathML types", function() {
-        var identifier = getMathML("x").children[0].children[0];
-        expect(identifier.children[0].type).toEqual("mi");
-
-        var number = getMathML("1").children[0].children[0];
-        expect(number.children[0].type).toEqual("mn");
-
-        var operator = getMathML("+").children[0].children[0];
-        expect(operator.children[0].type).toEqual("mo");
-
-        var space = getMathML("\\;").children[0].children[0];
-        expect(space.children[0].type).toEqual("mspace");
-
-        var text = getMathML("\\text{a}").children[0].children[0];
-        expect(text.children[0].type).toEqual("mtext");
-
-        var textop = getMathML("\\sin").children[0].children[0];
-        expect(textop.children[0].type).toEqual("mi");
-    });
-
-    it("should generate a <mphantom> node for \\phantom", function() {
-        var phantom = getMathML("\\phantom{x}").children[0].children[0];
-        expect(phantom.children[0].type).toEqual("mphantom");
-    });
-});
-
 describe("A parser that does not throw on unsupported commands", function() {
     // The parser breaks on unsupported commands unless it is explicitly
     // told not to
-    var errorColor = "#933";
-    var noThrowSettings = new Settings({
+    const errorColor = "#933";
+    const noThrowSettings = new Settings({
         throwOnError: false,
         errorColor: errorColor,
     });
@@ -1843,7 +2364,7 @@ describe("A parser that does not throw on unsupported commands", function() {
     });
 
     it("should produce color nodes with a color value given by errorColor", function() {
-        var parsedInput = getParsed("\\error", noThrowSettings);
+        const parsedInput = getParsed("\\error", noThrowSettings);
         expect(parsedInput[0].type).toBe("color");
         expect(parsedInput[0].value.color).toBe(errorColor);
     });
@@ -1860,8 +2381,8 @@ describe("The symbol table integraty", function() {
 
 describe("A macro expander", function() {
 
-    var compareParseTree = function(actual, expected, macros) {
-        var settings = new Settings({macros: macros});
+    const compareParseTree = function(actual, expected, macros) {
+        const settings = new Settings({macros: macros});
         actual = stripPositions(parseTree(actual, settings));
         expected = stripPositions(parseTree(expected, defaultSettings));
         expect(actual).toEqual(expected);
@@ -1871,10 +2392,123 @@ describe("A macro expander", function() {
         compareParseTree("e^\\foo", "e^1 23", {"\\foo": "123"});
     });
 
+    it("should preserve leading spaces inside macro definition", function() {
+        compareParseTree("\\text{\\foo}", "\\text{ x}", {"\\foo": " x"});
+    });
+
+    it("should preserve leading spaces inside macro argument", function() {
+        compareParseTree("\\text{\\foo{ x}}", "\\text{ x}", {"\\foo": "#1"});
+    });
+
+    it("should ignore expanded spaces in math mode", function() {
+        compareParseTree("\\foo", "x", {"\\foo": " x"});
+    });
+
+    it("should consume spaces after macro", function() {
+        compareParseTree("\\text{\\foo }", "\\text{x}", {"\\foo": "x"});
+    });
+
+    it("should consume spaces between arguments", function() {
+        compareParseTree("\\text{\\foo 1 2}", "\\text{12end}", {"\\foo": "#1#2end"});
+        compareParseTree("\\text{\\foo {1} {2}}", "\\text{12end}", {"\\foo": "#1#2end"});
+    });
+
     it("should allow for multiple expansion", function() {
         compareParseTree("1\\foo2", "1aa2", {
             "\\foo": "\\bar\\bar",
             "\\bar": "a",
         });
+    });
+
+    it("should allow for multiple expansion with argument", function() {
+        compareParseTree("1\\foo2", "12222", {
+            "\\foo": "\\bar{#1}\\bar{#1}",
+            "\\bar": "#1#1",
+        });
+    });
+
+    it("should allow for macro argument", function() {
+        compareParseTree("\\foo\\bar", "(x)", {
+            "\\foo": "(#1)",
+            "\\bar": "x",
+        });
+    });
+
+    it("should allow for space macro argument (text version)", function() {
+        compareParseTree("\\text{\\foo\\bar}", "\\text{( )}", {
+            "\\foo": "(#1)",
+            "\\bar": " ",
+        });
+    });
+
+    it("should allow for space macro argument (math version)", function() {
+        compareParseTree("\\foo\\bar", "()", {
+            "\\foo": "(#1)",
+            "\\bar": " ",
+        });
+    });
+
+    it("should allow for empty macro argument", function() {
+        compareParseTree("\\foo\\bar", "()", {
+            "\\foo": "(#1)",
+            "\\bar": "",
+        });
+    });
+
+    it("should expand the \\overset macro as expected", function() {
+        expect("\\overset?=").toParseLike("\\mathop{=}\\limits^{?}");
+        expect("\\overset{x=y}{\\sqrt{ab}}")
+            .toParseLike("\\mathop{\\sqrt{ab}}\\limits^{x=y}");
+        expect("\\overset {?} =").toParseLike("\\mathop{=}\\limits^{?}");
+    });
+
+    it("should build \\iff, \\implies, \\impliedby", function() {
+        expect("X \\iff Y").toBuild();
+        expect("X \\implies Y").toBuild();
+        expect("X \\impliedby Y").toBuild();
+    });
+});
+
+describe("A parser taking String objects", function() {
+    it("should not fail on an empty String object", function() {
+        expect(new String("")).toParse();
+    });
+
+    it("should parse the same as a regular string", function() {
+        expect(new String("xy")).toParseLike("xy");
+        expect(new String("\\div")).toParseLike("\\div");
+        expect(new String("\\frac 1 2")).toParseLike("\\frac 1 2");
+    });
+});
+
+describe("Unicode", function() {
+    it("should parse all lower case Greek letters", function() {
+        expect("αβγδεϵζηθϑικλμνξοπϖρϱςστυφϕχψω").toParse();
+    });
+
+    it("should parse 'ΓΔΘΞΠΣΦΨΩ'", function() {
+        expect("ΓΔΘΞΠΣΦΨΩ").toParse();
+    });
+});
+
+describe("The maxSize setting", function() {
+    const rule = "\\rule{999em}{999em}";
+
+    it("should clamp size when set", function() {
+        const built = getBuilt(rule, new Settings({maxSize: 5}))[0];
+        expect(built.style.borderRightWidth).toEqual("5em");
+        expect(built.style.borderTopWidth).toEqual("5em");
+    });
+
+    it("should not clamp size when not set", function() {
+        const built = getBuilt(rule)[0];
+        expect(built.style.borderRightWidth).toEqual("999em");
+        expect(built.style.borderTopWidth).toEqual("999em");
+    });
+
+    it("should make zero-width rules if a negative maxSize is passed", function() {
+        const built = getBuilt(rule, new Settings({maxSize: -5}))[0];
+        expect(built.style.borderRightWidth).toEqual("0em");
+        expect(built.style.borderTopWidth).toEqual("0em");
     });
 });
