@@ -1,12 +1,12 @@
 /* eslint no-console:0 */
 /* global katex */
 
-var splitAtDelimiters = require("./splitAtDelimiters");
+const splitAtDelimiters = require("./splitAtDelimiters");
 
-var splitWithDelimiters = function(text, delimiters) {
-    var data = [{type: "text", data: text}];
-    for (var i = 0; i < delimiters.length; i++) {
-        var delimiter = delimiters[i];
+const splitWithDelimiters = function(text, delimiters) {
+    let data = [{type: "text", data: text}];
+    for (let i = 0; i < delimiters.length; i++) {
+        const delimiter = delimiters[i];
         data = splitAtDelimiters(
             data, delimiter.left, delimiter.right,
             delimiter.display || false);
@@ -14,26 +14,29 @@ var splitWithDelimiters = function(text, delimiters) {
     return data;
 };
 
-var renderMathInText = function(text, delimiters) {
-    var data = splitWithDelimiters(text, delimiters);
+/* Note: optionsCopy is mutated by this method. If it is ever exposed in the
+ * API, we should copy it before mutating.
+ */
+const renderMathInText = function(text, optionsCopy) {
+    const data = splitWithDelimiters(text, optionsCopy.delimiters);
+    const fragment = document.createDocumentFragment();
 
-    var fragment = document.createDocumentFragment();
-
-    for (var i = 0; i < data.length; i++) {
+    for (let i = 0; i < data.length; i++) {
         if (data[i].type === "text") {
             fragment.appendChild(document.createTextNode(data[i].data));
         } else {
-            var span = document.createElement("span");
-            var math = data[i].data;
+            const span = document.createElement("span");
+            const math = data[i].data;
+            // Override any display mode defined in the settings with that
+            // defined by the text itself
+            optionsCopy.displayMode = data[i].display;
             try {
-                katex.render(math, span, {
-                    displayMode: data[i].display,
-                });
+                katex.render(math, span, optionsCopy);
             } catch (e) {
                 if (!(e instanceof katex.ParseError)) {
                     throw e;
                 }
-                console.error(
+                optionsCopy.errorCallback(
                     "KaTeX auto-render: Failed to parse `" + data[i].data +
                     "` with ",
                     e
@@ -48,28 +51,28 @@ var renderMathInText = function(text, delimiters) {
     return fragment;
 };
 
-var renderElem = function(elem, delimiters, ignoredTags) {
-    for (var i = 0; i < elem.childNodes.length; i++) {
-        var childNode = elem.childNodes[i];
+const renderElem = function(elem, optionsCopy) {
+    for (let i = 0; i < elem.childNodes.length; i++) {
+        const childNode = elem.childNodes[i];
         if (childNode.nodeType === 3) {
             // Text node
-            var frag = renderMathInText(childNode.textContent, delimiters);
+            const frag = renderMathInText(childNode.textContent, optionsCopy);
             i += frag.childNodes.length - 1;
             elem.replaceChild(frag, childNode);
         } else if (childNode.nodeType === 1) {
             // Element node
-            var shouldRender = ignoredTags.indexOf(
+            const shouldRender = optionsCopy.ignoredTags.indexOf(
                 childNode.nodeName.toLowerCase()) === -1;
 
             if (shouldRender) {
-                renderElem(childNode, delimiters, ignoredTags);
+                renderElem(childNode, optionsCopy);
             }
         }
         // Otherwise, it's something else, and ignore it.
     }
 };
 
-var defaultOptions = {
+const defaultAutoRenderOptions = {
     delimiters: [
         {left: "$$", right: "$$", display: true},
         {left: "\\[", right: "\\]", display: true},
@@ -81,31 +84,19 @@ var defaultOptions = {
     ignoredTags: [
         "script", "noscript", "style", "textarea", "pre", "code",
     ],
+
+    errorCallback: function(msg, err) {
+        console.error(msg, err);
+    },
 };
 
-var extend = function(obj) {
-    // Adapted from underscore.js' `_.extend`. See LICENSE.txt for license.
-    var source;
-    var prop;
-    for (var i = 1, length = arguments.length; i < length; i++) {
-        source = arguments[i];
-        for (prop in source) {
-            if (Object.prototype.hasOwnProperty.call(source, prop)) {
-                obj[prop] = source[prop];
-            }
-        }
-    }
-    return obj;
-};
-
-var renderMathInElement = function(elem, options) {
+const renderMathInElement = function(elem, options) {
     if (!elem) {
         throw new Error("No element provided to render");
     }
 
-    options = extend({}, defaultOptions, options);
-
-    renderElem(elem, options.delimiters, options.ignoredTags);
+    const optionsCopy = Object.assign({}, defaultAutoRenderOptions, options);
+    renderElem(elem, optionsCopy);
 };
 
 module.exports = renderMathInElement;
