@@ -1,22 +1,23 @@
-/* eslint no-console:0 */
+/* eslint-env node, es6 */
+/* eslint-disable no-console */
 "use strict";
 
-var childProcess = require("child_process");
-var fs = require("fs");
-var path = require("path");
-var Q = require("q"); // To debug, pass Q_DEBUG=1 in the environment
-var pngparse = require("pngparse");
-var fft = require("ndarray-fft");
-var ndarray = require("ndarray-fft/node_modules/ndarray");
+const childProcess = require("child_process");
+const fs = require("fs");
+const path = require("path");
+const Q = require("q"); // To debug, pass Q_DEBUG=1 in the environment
+const pngparse = require("pngparse");
+const fft = require("ndarray-fft");
+const ndarray = require("ndarray");
 
-var data = require("../../test/screenshotter/ss_data");
+const data = require("../../test/screenshotter/ss_data");
 
 // Adapt node functions to Q promises
-var readFile = Q.denodeify(fs.readFile);
-var writeFile = Q.denodeify(fs.writeFile);
-var mkdir = Q.denodeify(fs.mkdir);
+const readFile = Q.denodeify(fs.readFile);
+const writeFile = Q.denodeify(fs.writeFile);
+const mkdir = Q.denodeify(fs.mkdir);
 
-var todo;
+let todo;
 if (process.argv.length > 2) {
     todo = process.argv.slice(2);
 } else {
@@ -26,23 +27,23 @@ if (process.argv.length > 2) {
 }
 
 // Dimensions used when we do the FFT-based alignment computation
-var alignWidth = 2048; // should be at least twice the width resp. height
-var alignHeight = 2048; // of the screenshots, and a power of two.
+const alignWidth = 2048; // should be at least twice the width resp. height
+const alignHeight = 2048; // of the screenshots, and a power of two.
 
 // Compute required resolution to match test.html. 16px default font,
 // scaled to 4em in test.html, and to 1.21em in katex.css. Corresponding
 // LaTeX font size is 10pt. There are 72.27pt per inch.
-var pxPerEm = 16 * 4 * 1.21;
-var pxPerPt = pxPerEm / 10;
-var dpi = pxPerPt * 72.27;
+const pxPerEm = 16 * 4 * 1.21;
+const pxPerPt = pxPerEm / 10;
+const dpi = pxPerPt * 72.27;
 
-var tmpDir = "/tmp/texcmp";
-var ssDir = path.normalize(
+const tmpDir = "/tmp/texcmp";
+const ssDir = path.normalize(
     path.join(__dirname, "..", "..", "test", "screenshotter"));
-var imagesDir = path.join(ssDir, "images");
-var teximgDir = path.join(ssDir, "tex");
-var diffDir = path.join(ssDir, "diff");
-var template;
+const imagesDir = path.join(ssDir, "images");
+const teximgDir = path.join(ssDir, "tex");
+const diffDir = path.join(ssDir, "diff");
+let template;
 
 Q.all([
     readFile(path.join(ssDir, "test.tex"), "utf-8"),
@@ -57,8 +58,8 @@ Q.all([
 
 // Process a single test case: rasterize, then create diff
 function processTestCase(key) {
-    var itm = data[key];
-    var tex = "$" + itm.tex + "$";
+    const itm = data[key];
+    let tex = "$" + itm.tex + "$";
     if (itm.display) {
         tex = "\\[" + itm.tex + "\\]";
     }
@@ -68,15 +69,32 @@ function processTestCase(key) {
     if (itm.post) {
         tex = tex + itm.post.replace("<br>", "\\\\");
     }
+    if (itm.macros) {
+        tex = Object.keys(itm.macros).map(name => {
+            const expansion = itm.macros[name];
+            let numArgs = 0;
+            if (expansion.indexOf("#") !== -1) {
+                const stripped = expansion.replace(/##/g, "");
+                while (stripped.indexOf("#" + (numArgs + 1)) !== -1) {
+                    ++numArgs;
+                }
+            }
+            let args = "";
+            for (let i = 1; i <= numArgs; ++i) {
+                args += "#" + i;
+            }
+            return "\\def" + name + args + "{" + expansion + "}\n";
+        }).join("") + tex;
+    }
     tex = template.replace(/\$.*\$/, tex.replace(/\$/g, "$$$$"));
-    var texFile = path.join(tmpDir, key + ".tex");
-    var pdfFile = path.join(tmpDir, key + ".pdf");
-    var pngFile = path.join(teximgDir, key + "-pdflatex.png");
-    var browserFile = path.join(imagesDir, key + "-firefox.png");
-    var diffFile = path.join(diffDir, key + ".png");
+    const texFile = path.join(tmpDir, key + ".tex");
+    const pdfFile = path.join(tmpDir, key + ".pdf");
+    const pngFile = path.join(teximgDir, key + "-pdflatex.png");
+    const browserFile = path.join(imagesDir, key + "-firefox.png");
+    const diffFile = path.join(diffDir, key + ".png");
 
     // Step 1: write key.tex file
-    var fftLatex = writeFile(texFile, tex).then(function() {
+    const fftLatex = writeFile(texFile, tex).then(function() {
         // Step 2: call "pdflatex key" to create key.pdf
         return execFile("pdflatex", [
             "-interaction", "nonstopmode", key,
@@ -86,7 +104,7 @@ function processTestCase(key) {
         // Step 3: call "convert ... key.pdf key.png" to create key.png
         return execFile("convert", [
             "-density", dpi, "-units", "PixelsPerInch", "-flatten",
-            pdfFile, pngFile,
+            "-depth", "8", pdfFile, pngFile,
         ]);
     }).then(function() {
         console.log("Rasterized " + key);
@@ -94,24 +112,24 @@ function processTestCase(key) {
         return readPNG(pngFile).then(fftImage);
     });
     // Step 5: apply FFT to reference image as well
-    var fftBrowser = readPNG(browserFile).then(fftImage);
+    const fftBrowser = readPNG(browserFile).then(fftImage);
 
     return Q.all([fftBrowser, fftLatex]).spread(function(browser, latex) {
         // Now we have the FFT result from both
         // Step 6: find alignment which maximizes overlap.
         // This uses a FFT-based correlation computation.
-        var x;
-        var y;
-        var real = createMatrix();
-        var imag = createMatrix();
+        let x;
+        let y;
+        const real = createMatrix();
+        const imag = createMatrix();
 
         // Step 6a: (real + i*imag) = latex * conjugate(browser)
         for (y = 0; y < alignHeight; ++y) {
             for (x = 0; x < alignWidth; ++x) {
-                var br = browser.real.get(y, x);
-                var bi = browser.imag.get(y, x);
-                var lr = latex.real.get(y, x);
-                var li = latex.imag.get(y, x);
+                const br = browser.real.get(y, x);
+                const bi = browser.imag.get(y, x);
+                const lr = latex.real.get(y, x);
+                const li = latex.imag.get(y, x);
                 real.set(y, x, br * lr + bi * li);
                 imag.set(y, x, br * li - bi * lr);
             }
@@ -121,14 +139,14 @@ function processTestCase(key) {
         fft(-1, real, imag);
 
         // Step 6c: find position where the (squared) absolute value is maximal
-        var offsetX = 0;
-        var offsetY = 0;
-        var maxSquaredNorm = -1; // any result is greater than initial value
+        let offsetX = 0;
+        let offsetY = 0;
+        let maxSquaredNorm = -1; // any result is greater than initial value
         for (y = 0; y < alignHeight; ++y) {
             for (x = 0; x < alignWidth; ++x) {
-                var or = real.get(y, x);
-                var oi = imag.get(y, x);
-                var squaredNorm = or * or + oi * oi;
+                const or = real.get(y, x);
+                const oi = imag.get(y, x);
+                const squaredNorm = or * or + oi * oi;
                 if (maxSquaredNorm < squaredNorm) {
                     maxSquaredNorm = squaredNorm;
                     offsetX = x;
@@ -147,19 +165,19 @@ function processTestCase(key) {
         console.log("Positioned " + key + ": " + offsetX + ", " + offsetY);
 
         // Step 7: use these offsets to compute difference illustration
-        var bx = Math.max(offsetX, 0); // browser left padding
-        var by = Math.max(offsetY, 0); // browser top padding
-        var lx = Math.max(-offsetX, 0); // latex left padding
-        var ly = Math.max(-offsetY, 0); // latex top padding
-        var uw = Math.max(browser.width + bx, latex.width + lx); // union width
-        var uh = Math.max(browser.height + by, latex.height + ly); // u. height
+        const bx = Math.max(offsetX, 0); // browser left padding
+        const by = Math.max(offsetY, 0); // browser top padding
+        const lx = Math.max(-offsetX, 0); // latex left padding
+        const ly = Math.max(-offsetY, 0); // latex top padding
+        const uw = Math.max(browser.width + bx, latex.width + lx); // union w.
+        const uh = Math.max(browser.height + by, latex.height + ly); // u. h.
         return execFile("convert", [
             // First image: latex rendering, converted to grayscale and padded
-            "(", pngFile, "-grayscale", "Rec709Luminance",
+            "(", pngFile, "-colorspace", "Gray",
             "-extent", uw + "x" + uh + "-" + lx + "-" + ly,
             ")",
             // Second image: browser screenshot, to grayscale and padded
-            "(", browserFile, "-grayscale", "Rec709Luminance",
+            "(", browserFile, "-colorspace", "Gray",
             "-extent", uw + "x" + uh + "-" + bx + "-" + by,
             ")",
             // Third image: the per-pixel minimum of the first two images
@@ -186,7 +204,7 @@ function ensureDir(dir) {
 // Execute a given command, and return a promise to its output.
 // Don't denodeify here, since fail branch needs access to stderr.
 function execFile(cmd, args, opts) {
-    var deferred = Q.defer();
+    const deferred = Q.defer();
     childProcess.execFile(cmd, args, opts, function(err, stdout, stderr) {
         if (err) {
             console.error("Error executing " + cmd + " " + args.join(" "));
@@ -203,12 +221,13 @@ function execFile(cmd, args, opts) {
 
 // Read given file and parse it as a PNG file.
 function readPNG(file) {
-    var deferred = Q.defer();
-    var onerror = deferred.reject.bind(deferred);
-    var stream = fs.createReadStream(file);
+    const deferred = Q.defer();
+    const onerror = deferred.reject.bind(deferred);
+    const stream = fs.createReadStream(file);
     stream.on("error", onerror);
     pngparse.parseStream(stream, function(err, image) {
         if (err) {
+            console.log("Failed to load " + file);
             onerror(err);
             return;
         }
@@ -219,20 +238,19 @@ function readPNG(file) {
 
 // Take a parsed image data structure and apply FFT transformation to it
 function fftImage(image) {
-    var real = createMatrix();
-    var imag = createMatrix();
-    var idx = 0;
-    var nchan = image.channels;
-    var alphachan = 1 - (nchan % 2);
-    var colorchan = nchan - alphachan;
-    for (var y = 0; y < image.height; ++y) {
-        for (var x = 0; x < image.width; ++x) {
-            var c;
-            var v = 0;
-            for (c = 0; c < colorchan; ++c) {
+    const real = createMatrix();
+    const imag = createMatrix();
+    let idx = 0;
+    const nchan = image.channels;
+    const alphachan = 1 - (nchan % 2);
+    const colorchan = nchan - alphachan;
+    for (let y = 0; y < image.height; ++y) {
+        for (let x = 0; x < image.width; ++x) {
+            let v = 0;
+            for (let c = 0; c < colorchan; ++c) {
                 v += 255 - image.data[idx++];
             }
-            for (c = 0; c < alphachan; ++c) {
+            for (let c = 0; c < alphachan; ++c) {
                 v += image.data[idx++];
             }
             real.set(y, x, v);
@@ -249,6 +267,6 @@ function fftImage(image) {
 
 // Create a new matrix of preconfigured dimensions, initialized to zero
 function createMatrix() {
-    var array = new Float64Array(alignWidth * alignHeight);
+    const array = new Float64Array(alignWidth * alignHeight);
     return new ndarray(array, [alignWidth, alignHeight]);
 }
