@@ -1,37 +1,41 @@
 /* eslint no-console:0 */
-var fs = require("fs");
-var path = require("path");
+const fs = require("fs");
+const path = require("path");
 
-var browserify = require("browserify");
-var express = require("express");
-var glob = require("glob");
-var less = require("less");
+const babelify = require("babelify");
+const browserify = require("browserify");
+const express = require("express");
+const glob = require("glob");
+const less = require("less");
 
-var app = express();
+const app = express();
 
 if (require.main === module) {
-    app.use(express.logger());
+    app.use(require("morgan")(
+        ":date[iso] :method :url HTTP/:http-version - :status"));
 }
 
-var serveBrowserified = function(file, standaloneName) {
+function serveBrowserified(file, standaloneName) {
     return function(req, res, next) {
-        var files;
+        let files;
         if (Array.isArray(file)) {
-            files = file;
+            files = file.map(function(f) { return path.join(__dirname, f); });
         } else if (file.indexOf("*") !== -1) {
-            files = glob.sync(file);
+            files = glob.sync(file, {cwd: __dirname});
         } else {
-            files = [file];
+            files = [path.join(__dirname, file)];
         }
 
-        var options = {};
+        const options = {
+            transform: [babelify],
+        };
         if (standaloneName) {
             options.standalone = standaloneName;
         }
-        var b = browserify(files, options);
-        var stream = b.bundle();
+        const b = browserify(files, options);
+        const stream = b.bundle();
 
-        var body = "";
+        let body = "";
         stream.on("data", function(s) { body += s; });
         stream.on("error", function(e) { next(e); });
         stream.on("end", function() {
@@ -39,30 +43,36 @@ var serveBrowserified = function(file, standaloneName) {
             res.send(body);
         });
     };
-};
+}
 
-app.get("/katex.js", serveBrowserified("./katex", "katex"));
-app.use("/test/jasmine",
-    express["static"](
-        path.dirname(
-            require.resolve("jasmine-core/lib/jasmine-core/jasmine.js")
-        )
-    )
-);
-app.get("/test/katex-spec.js", serveBrowserified("./test/*[Ss]pec.js"));
-app.get("/contrib/auto-render/auto-render.js",
-        serveBrowserified("./contrib/auto-render/auto-render",
-                          "renderMathInElement"));
+function browserified(url, file, standaloneName) {
+    app.get(url, serveBrowserified(file, standaloneName));
+}
 
-app.get("/katex.css", function(req, res, next) {
-    fs.readFile("static/katex.less", {encoding: "utf8"}, function(err, data) {
+function getStatic(url, file) {
+    app.use(url, express.static(path.join(__dirname, file)));
+}
+
+browserified("/katex.js", "katex", "katex");
+browserified("/test/katex-spec.js", "test/*[Ss]pec.js");
+browserified(
+    "/contrib/auto-render/auto-render.js",
+    "contrib/auto-render/auto-render",
+    "renderMathInElement");
+browserified(
+    "/contrib/copy-tex/copy-tex.js",
+    "contrib/copy-tex/copy-tex");
+
+app.use("/katex.css", function(req, res, next) {
+    const lessfile = path.join(__dirname, "static", "katex.less");
+    fs.readFile(lessfile, {encoding: "utf8"}, function(err, data) {
         if (err) {
             next(err);
             return;
         }
 
         less.render(data, {
-            paths: ["./static"],
+            paths: [path.join(__dirname, "static")],
             filename: "katex.less",
         }, function(err, output) {
             if (err) {
@@ -77,10 +87,10 @@ app.get("/katex.css", function(req, res, next) {
     });
 });
 
-app.use(express["static"](path.join(__dirname, "static")));
-app.use(express["static"](path.join(__dirname, "build")));
-app.use("/test", express["static"](path.join(__dirname, "test")));
-app.use("/contrib", express["static"](path.join(__dirname, "contrib")));
+getStatic("", "static");
+getStatic("", "build");
+getStatic("/test", "test");
+getStatic("/contrib", "contrib");
 
 app.use(function(err, req, res, next) {
     console.error(err.stack);
