@@ -1,4 +1,5 @@
-module.exports = {
+// @flow
+const fontMetricsData = {
     "AMS-Regular": {
         "65": [0, 0.68889, 0, 0],
         "66": [0, 0.68889, 0, 0],
@@ -1750,3 +1751,5 @@ module.exports = {
         "8242": [0, 0.61111, 0, 0],
     },
 };
+
+module.exports = fontMetricsData;
