@@ -5,7 +5,7 @@ function init() {
 
     if ("oninput" in input) {
         input.addEventListener("input", reprocess, false);
-    } else {
+    } else if (input.attachEvent) {
         input.attachEvent("onkeyup", reprocess);
     }
 
@@ -28,6 +28,30 @@ function init() {
         options.macros = macros;
         macros["\\" + match[1]] = decodeURIComponent(match[2]);
         macroString += "&" + match[0].substr(1);
+    }
+
+    // The `before` search parameter puts normal text before the math.
+    // The `after` search parameter puts normal text after the math.
+    // Example use: testing baseline alignment.
+    if (/(?:^\?|&)(?:before|after)=/.test(window.location.search)) {
+        var mathContainer = math;
+        mathContainer.id = "math-container";
+
+        if ((match = /(?:^\?|&)before=([^&]*)/.exec(window.location.search))) {
+            var child = document.createTextNode(decodeURIComponent(match[1]));
+            mathContainer.appendChild(child);
+            macroString += "&" + match[0].substr(1);
+        }
+
+        math = document.createElement("span");
+        math.id = "math";
+        mathContainer.appendChild(math);
+
+        if ((match = /(?:^\?|&)after=([^&]*)/.exec(window.location.search))) {
+            var child = document.createTextNode(decodeURIComponent(match[1]));
+            mathContainer.appendChild(child);
+            macroString += "&" + match[0].substr(1);
+        }
     }
 
     reprocess();
